Make Jest work, update / upgrade dependencies, remove unnecessary dependencies
