+++ package.json
@@ -35,13 +35,16 @@
     "module": "index.js",
     "typings": "index.d.ts",
     "scripts": {
-        "webpack-dev-server": "webpack-dev-server",
         "gulp": "gulp",
+        "webpack-dev-server": "webpack-dev-server",
 		"build:dev": "webpack --config ./tools/webpack/webpack.config.js --hide-modules",
+        "build:publish": "gulp all",
 		"start": "webpack --config ./tools/webpack/webpack.dll.config.js --hide-modules && webpack-dev-server --config ./tools/webpack/webpack.dev.config.js",
         "release:automatic-release": "automatic-release",
         "release:list-deployed-files": "pkgfiles --sort=name",
-        "coveralls": "cat ./test-coverage/ts/lcov.info | coveralls"
+        "coveralls": "cat ./coverage/remapped/lcov.info | coveralls",
+        "test": "jest --config ./tools/jest/jest.config.json",
+		"test:ci": "jest --config ./tools/jest/jest.config.json --runInBand"
     },
     "peerDependencies": {
         "@angular/core": ">= 4.0.0 < 5.0.0",
@@ -54,63 +57,51 @@
         "@angular/compiler-cli": "4.0.x",
         "@angular/platform-browser": "4.0.x",
         "@angular/platform-browser-dynamic": "4.0.x",
-        "@types/jasmine": "2.5.x",
+        "@types/jest": "19.2.x",
         "@types/node": "7.0.x",
-        "add-asset-html-webpack-plugin": "1.0.x",
+        "add-asset-html-webpack-plugin": "2.0.x",
         "angular2-template-loader": "0.6.x",
         "automatic-release": "1.0.x",
         "awesome-typescript-loader": "3.1.x",
         "browser-sync": "2.18.x",
         "browser-sync-webpack-plugin": "1.1.x",
-        "chalk": "1.1.x",
         "classlist.js": "1.1.x",
-        "codelyzer": "2.0.0-beta.4",
+        "codelyzer": "3.0.x",
         "core-js": "2.4.x",
-        "coveralls": "2.11.x",
+        "coveralls": "2.13.x",
         "css-loader": "0.28.x",
         "del": "2.2.x",
         "friendly-errors-webpack-plugin": "1.6.x",
         "gulp": "github:gulpjs/gulp#4.0",
         "gulp-autoprefixer": "3.1.x",
-        "gulp-clean-css": "2.3.x",
         "gulp-inline-ng2-template": "4.0.x",
         "gulp-ngc": "0.2.x",
         "gulp-rename": "1.2.x",
         "gulp-rollup": "2.11.x",
-        "gulp-sourcemaps": "2.4.x",
+        "gulp-sourcemaps": "2.6.x",
         "gulp-sass": "3.1.x",
         "gulp-sass-lint": "1.3.x",
-        "gulp-tslint": "7.0.x",
+        "gulp-tslint": "8.0.x",
         "gulp-typescript": "3.1.x",
-        "gulp-util": "3.0.x",
-        "html-minifier": "3.3.x",
+        "html-minifier": "3.4.x",
         "html-webpack-plugin": "2.28.x",
-        "jasmine-core": "2.5.x",
-        "karma": "1.4.x",
-        "karma-chrome-launcher": "2.0.x",
-        "karma-cli": "1.0.x",
-        "karma-coverage": "1.1.x",
-        "karma-jasmine": "1.1.x",
-        "karma-mocha-reporter": "2.2.x",
-        "merge2": "1.0.x",
+        "jest": "19.0.x",
+		"jest-preset-angular": "1.1.x",
         "pkgfiles": "2.3.x",
         "progress-bar-webpack-plugin": "1.9.x",
         "raw-loader": "0.5.x",
         "reflect-metadata": "0.1.x",
-        "remap-istanbul": "0.8.x",
-        "require-dir": "0.3.x",
         "rxjs": "5.3.x",
         "sass-loader": "6.0.x",
         "script-ext-html-webpack-plugin": "1.7.x",
         "simple-progress-webpack-plugin": "1.0.x",
-        "source-map-loader": "0.1.x",
+        "source-map-loader": "0.2.x",
         "style-loader": "0.16.x",
-        "systemjs": "0.20.x",
-        "tslint": "4.4.x",
+        "tslint": "5.1.x",
         "typescript": "2.2.x",
         "web-animations-js": "2.2.x",
         "webpack": "2.4.x",
         "webpack-dev-server": "2.4.x",
-        "zone.js": "0.8.5"
+        "zone.js": "0.8.8"
     }
 }
