+++ src/lib/src/animation-presets/slide.animation-preset.spec.ts
@@ -5,269 +5,265 @@ import { slide } from './../animation-presets/slide.animation-preset';
 /**
  * Slide Animation Preset - Unit Test
  */
-export function main(): void {
+describe( 'Slide Animation Preset', () => {
 
-	describe( 'Slide Animation Preset', () => {
+	describe( '(show)', () => {
 
-		describe( '(show)', () => {
+		it( 'should return animation keyframes for top-left position', () => {
 
-			it( 'should return animation keyframes for top-left position', () => {
-
-				const testConfig: NotifierConfig = new NotifierConfig( {
-					position: {
-						horizontal: {
-							distance: 50,
-							position: 'left'
-						},
-						vertical: {
-							distance: 100,
-							position: 'top'
-						}
-					}
-				} );
-				const testNotification: MockNotification = new MockNotification( testConfig );
-				const expectedKeyframes: NotifierAnimationPresetKeyframes = {
-					from: {
-						transform: `translate3d( calc( -100% - ${ testConfig.position.horizontal.distance }px - 10px ), 0, 0 )`
+			const testConfig: NotifierConfig = new NotifierConfig( {
+				position: {
+					horizontal: {
+						distance: 50,
+						position: 'left'
 					},
-					to: {
-						transform: 'translate3d( 0, 0, 0 )'
+					vertical: {
+						distance: 100,
+						position: 'top'
 					}
-				};
-				// tslint:disable no-any
-				const keyframes: NotifierAnimationPresetKeyframes = slide.show( <any> testNotification );
-				// tslint:enable no-any
-
-				expect( keyframes ).toEqual( expectedKeyframes );
-
+				}
 			} );
+			const testNotification: MockNotification = new MockNotification( testConfig );
+			const expectedKeyframes: NotifierAnimationPresetKeyframes = {
+				from: {
+					transform: `translate3d( calc( -100% - ${ testConfig.position.horizontal.distance }px - 10px ), 0, 0 )`
+				},
+				to: {
+					transform: 'translate3d( 0, 0, 0 )'
+				}
+			};
+			// tslint:disable no-any
+			const keyframes: NotifierAnimationPresetKeyframes = slide.show( <any> testNotification );
+			// tslint:enable no-any
 
-			it( 'should return animation keyframes for top-right position', () => {
-
-				const testConfig: NotifierConfig = new NotifierConfig( {
-					position: {
-						horizontal: {
-							distance: 50,
-							position: 'right'
-						},
-						vertical: {
-							distance: 100,
-							position: 'top'
-						}
-					}
-				} );
-				const testNotification: MockNotification = new MockNotification( testConfig );
-				const expectedKeyframes: NotifierAnimationPresetKeyframes = {
-					from: {
-						transform: `translate3d( calc( 100% + ${ testConfig.position.horizontal.distance }px + 10px ), 0, 0 )`
-					},
-					to: {
-						transform: 'translate3d( 0, 0, 0 )'
-					}
-				};
-				// tslint:disable no-any
-				const keyframes: NotifierAnimationPresetKeyframes = slide.show( <any> testNotification );
-				// tslint:enable no-any
-
-				expect( keyframes ).toEqual( expectedKeyframes );
-
-			} );
-
-			it( 'should return animation keyframes for top-middle position', () => {
-
-				const testConfig: NotifierConfig = new NotifierConfig( {
-					position: {
-						horizontal: {
-							distance: 50,
-							position: 'middle'
-						},
-						vertical: {
-							distance: 100,
-							position: 'top'
-						}
-					}
-				} );
-				const testNotification: MockNotification = new MockNotification( testConfig );
-				const expectedKeyframes: NotifierAnimationPresetKeyframes = {
-					from: {
-						transform: `translate3d( -50%, calc( -100% - ${ testConfig.position.horizontal.distance }px - 10px ), 0 )`
-					},
-					to: {
-						transform: 'translate3d( -50%, 0, 0 )'
-					}
-				};
-				// tslint:disable no-any
-				const keyframes: NotifierAnimationPresetKeyframes = slide.show( <any> testNotification );
-				// tslint:enable no-any
-
-				expect( keyframes ).toEqual( expectedKeyframes );
-
-			} );
-
-			it( 'should return animation keyframes for bottom-middle position', () => {
-
-				const testConfig: NotifierConfig = new NotifierConfig( {
-					position: {
-						horizontal: {
-							distance: 50,
-							position: 'middle'
-						},
-						vertical: {
-							distance: 100,
-							position: 'bottom'
-						}
-					}
-				} );
-				const testNotification: MockNotification = new MockNotification( testConfig );
-				const expectedKeyframes: NotifierAnimationPresetKeyframes = {
-					from: {
-						transform: `translate3d( -50%, calc( 100% + ${ testConfig.position.horizontal.distance }px + 10px ), 0 )`
-					},
-					to: {
-						transform: 'translate3d( -50%, 0, 0 )'
-					}
-				};
-				// tslint:disable no-any
-				const keyframes: NotifierAnimationPresetKeyframes = slide.show( <any> testNotification );
-				// tslint:enable no-any
-
-				expect( keyframes ).toEqual( expectedKeyframes );
-
-			} );
+			expect( keyframes ).toEqual( expectedKeyframes );
 
 		} );
 
-		describe( '(hide)', () => {
+		it( 'should return animation keyframes for top-right position', () => {
 
-			it( 'should return animation keyframes for top-left position', () => {
-				const testConfig: NotifierConfig = new NotifierConfig( {
-					position: {
-						horizontal: {
-							distance: 50,
-							position: 'left'
-						},
-						vertical: {
-							distance: 100,
-							position: 'top'
-						}
-					}
-				} );
-				const testNotification: MockNotification = new MockNotification( testConfig );
-				const expectedKeyframes: NotifierAnimationPresetKeyframes = {
-					from: {
-						transform: `translate3d( 0, ${ testNotification.component.getShift() }px, 0 )`
+			const testConfig: NotifierConfig = new NotifierConfig( {
+				position: {
+					horizontal: {
+						distance: 50,
+						position: 'right'
 					},
-					to: {
-						/* tslint:disable max-line-length */
-						transform: `translate3d( calc( -100% - ${ testConfig.position.horizontal.distance }px - 10px ), ${ testNotification.component.getShift() }px, 0 )`
-						/* tslint:enable max-line-length */
+					vertical: {
+						distance: 100,
+						position: 'top'
 					}
-				};
-				// tslint:disable no-any
-				const keyframes: NotifierAnimationPresetKeyframes = slide.hide( <any> testNotification );
-				// tslint:enable no-any
-
-				expect( keyframes ).toEqual( expectedKeyframes );
-
+				}
 			} );
+			const testNotification: MockNotification = new MockNotification( testConfig );
+			const expectedKeyframes: NotifierAnimationPresetKeyframes = {
+				from: {
+					transform: `translate3d( calc( 100% + ${ testConfig.position.horizontal.distance }px + 10px ), 0, 0 )`
+				},
+				to: {
+					transform: 'translate3d( 0, 0, 0 )'
+				}
+			};
+			// tslint:disable no-any
+			const keyframes: NotifierAnimationPresetKeyframes = slide.show( <any> testNotification );
+			// tslint:enable no-any
 
-			it( 'should return animation keyframes for top-right position', () => {
-				const testConfig: NotifierConfig = new NotifierConfig( {
-					position: {
-						horizontal: {
-							distance: 50,
-							position: 'right'
-						},
-						vertical: {
-							distance: 100,
-							position: 'top'
-						}
-					}
-				} );
-				const testNotification: MockNotification = new MockNotification( testConfig );
-				const expectedKeyframes: NotifierAnimationPresetKeyframes = {
-					from: {
-						transform: `translate3d( 0, ${ testNotification.component.getShift() }px, 0 )`
+			expect( keyframes ).toEqual( expectedKeyframes );
+
+		} );
+
+		it( 'should return animation keyframes for top-middle position', () => {
+
+			const testConfig: NotifierConfig = new NotifierConfig( {
+				position: {
+					horizontal: {
+						distance: 50,
+						position: 'middle'
 					},
-					to: {
-						/* tslint:disable max-line-length */
-						transform: `translate3d( calc( 100% + ${ testConfig.position.horizontal.distance }px + 10px ), ${ testNotification.component.getShift() }px, 0 )`
-						/* tslint:enable max-line-length */
+					vertical: {
+						distance: 100,
+						position: 'top'
 					}
-				};
-				// tslint:disable no-any
-				const keyframes: NotifierAnimationPresetKeyframes = slide.hide( <any> testNotification );
-				// tslint:enable no-any
-
-				expect( keyframes ).toEqual( expectedKeyframes );
-
+				}
 			} );
+			const testNotification: MockNotification = new MockNotification( testConfig );
+			const expectedKeyframes: NotifierAnimationPresetKeyframes = {
+				from: {
+					transform: `translate3d( -50%, calc( -100% - ${ testConfig.position.horizontal.distance }px - 10px ), 0 )`
+				},
+				to: {
+					transform: 'translate3d( -50%, 0, 0 )'
+				}
+			};
+			// tslint:disable no-any
+			const keyframes: NotifierAnimationPresetKeyframes = slide.show( <any> testNotification );
+			// tslint:enable no-any
 
-			it( 'should return animation keyframes for top-middle position', () => {
-				const testConfig: NotifierConfig = new NotifierConfig( {
-					position: {
-						horizontal: {
-							distance: 50,
-							position: 'middle'
-						},
-						vertical: {
-							distance: 100,
-							position: 'top'
-						}
-					}
-				} );
-				const testNotification: MockNotification = new MockNotification( testConfig );
-				const expectedKeyframes: NotifierAnimationPresetKeyframes = {
-					from: {
-						transform: `translate3d( -50%, ${ testNotification.component.getShift() }px, 0 )`
+			expect( keyframes ).toEqual( expectedKeyframes );
+
+		} );
+
+		it( 'should return animation keyframes for bottom-middle position', () => {
+
+			const testConfig: NotifierConfig = new NotifierConfig( {
+				position: {
+					horizontal: {
+						distance: 50,
+						position: 'middle'
 					},
-					to: {
-						transform: `translate3d( -50%, calc( -100% - ${ testConfig.position.horizontal.distance }px - 10px ), 0 )`
+					vertical: {
+						distance: 100,
+						position: 'bottom'
 					}
-				};
-				// tslint:disable no-any
-				const keyframes: NotifierAnimationPresetKeyframes = slide.hide( <any> testNotification );
-				// tslint:enable no-any
-
-				expect( keyframes ).toEqual( expectedKeyframes );
-
+				}
 			} );
+			const testNotification: MockNotification = new MockNotification( testConfig );
+			const expectedKeyframes: NotifierAnimationPresetKeyframes = {
+				from: {
+					transform: `translate3d( -50%, calc( 100% + ${ testConfig.position.horizontal.distance }px + 10px ), 0 )`
+				},
+				to: {
+					transform: 'translate3d( -50%, 0, 0 )'
+				}
+			};
+			// tslint:disable no-any
+			const keyframes: NotifierAnimationPresetKeyframes = slide.show( <any> testNotification );
+			// tslint:enable no-any
 
-			it( 'should return animation keyframes for bottom-middle position', () => {
-				const testConfig: NotifierConfig = new NotifierConfig( {
-					position: {
-						horizontal: {
-							distance: 50,
-							position: 'middle'
-						},
-						vertical: {
-							distance: 100,
-							position: 'bottom'
-						}
-					}
-				} );
-				const testNotification: MockNotification = new MockNotification( testConfig );
-				const expectedKeyframes: NotifierAnimationPresetKeyframes = {
-					from: {
-						transform: `translate3d( -50%, ${ testNotification.component.getShift() }px, 0 )`
-					},
-					to: {
-						transform: `translate3d( -50%, calc( 100% + ${ testConfig.position.horizontal.distance }px + 10px ), 0 )`
-					}
-				};
-				// tslint:disable no-any
-				const keyframes: NotifierAnimationPresetKeyframes = slide.hide( <any> testNotification );
-				// tslint:enable no-any
-
-				expect( keyframes ).toEqual( expectedKeyframes );
-
-			} );
+			expect( keyframes ).toEqual( expectedKeyframes );
 
 		} );
 
 	} );
 
-}
+	describe( '(hide)', () => {
+
+		it( 'should return animation keyframes for top-left position', () => {
+			const testConfig: NotifierConfig = new NotifierConfig( {
+				position: {
+					horizontal: {
+						distance: 50,
+						position: 'left'
+					},
+					vertical: {
+						distance: 100,
+						position: 'top'
+					}
+				}
+			} );
+			const testNotification: MockNotification = new MockNotification( testConfig );
+			const expectedKeyframes: NotifierAnimationPresetKeyframes = {
+				from: {
+					transform: `translate3d( 0, ${ testNotification.component.getShift() }px, 0 )`
+				},
+				to: {
+					/* tslint:disable max-line-length */
+					transform: `translate3d( calc( -100% - ${ testConfig.position.horizontal.distance }px - 10px ), ${ testNotification.component.getShift() }px, 0 )`
+					/* tslint:enable max-line-length */
+				}
+			};
+			// tslint:disable no-any
+			const keyframes: NotifierAnimationPresetKeyframes = slide.hide( <any> testNotification );
+			// tslint:enable no-any
+
+			expect( keyframes ).toEqual( expectedKeyframes );
+
+		} );
+
+		it( 'should return animation keyframes for top-right position', () => {
+			const testConfig: NotifierConfig = new NotifierConfig( {
+				position: {
+					horizontal: {
+						distance: 50,
+						position: 'right'
+					},
+					vertical: {
+						distance: 100,
+						position: 'top'
+					}
+				}
+			} );
+			const testNotification: MockNotification = new MockNotification( testConfig );
+			const expectedKeyframes: NotifierAnimationPresetKeyframes = {
+				from: {
+					transform: `translate3d( 0, ${ testNotification.component.getShift() }px, 0 )`
+				},
+				to: {
+					/* tslint:disable max-line-length */
+					transform: `translate3d( calc( 100% + ${ testConfig.position.horizontal.distance }px + 10px ), ${ testNotification.component.getShift() }px, 0 )`
+					/* tslint:enable max-line-length */
+				}
+			};
+			// tslint:disable no-any
+			const keyframes: NotifierAnimationPresetKeyframes = slide.hide( <any> testNotification );
+			// tslint:enable no-any
+
+			expect( keyframes ).toEqual( expectedKeyframes );
+
+		} );
+
+		it( 'should return animation keyframes for top-middle position', () => {
+			const testConfig: NotifierConfig = new NotifierConfig( {
+				position: {
+					horizontal: {
+						distance: 50,
+						position: 'middle'
+					},
+					vertical: {
+						distance: 100,
+						position: 'top'
+					}
+				}
+			} );
+			const testNotification: MockNotification = new MockNotification( testConfig );
+			const expectedKeyframes: NotifierAnimationPresetKeyframes = {
+				from: {
+					transform: `translate3d( -50%, ${ testNotification.component.getShift() }px, 0 )`
+				},
+				to: {
+					transform: `translate3d( -50%, calc( -100% - ${ testConfig.position.horizontal.distance }px - 10px ), 0 )`
+				}
+			};
+			// tslint:disable no-any
+			const keyframes: NotifierAnimationPresetKeyframes = slide.hide( <any> testNotification );
+			// tslint:enable no-any
+
+			expect( keyframes ).toEqual( expectedKeyframes );
+
+		} );
+
+		it( 'should return animation keyframes for bottom-middle position', () => {
+			const testConfig: NotifierConfig = new NotifierConfig( {
+				position: {
+					horizontal: {
+						distance: 50,
+						position: 'middle'
+					},
+					vertical: {
+						distance: 100,
+						position: 'bottom'
+					}
+				}
+			} );
+			const testNotification: MockNotification = new MockNotification( testConfig );
+			const expectedKeyframes: NotifierAnimationPresetKeyframes = {
+				from: {
+					transform: `translate3d( -50%, ${ testNotification.component.getShift() }px, 0 )`
+				},
+				to: {
+					transform: `translate3d( -50%, calc( 100% + ${ testConfig.position.horizontal.distance }px + 10px ), 0 )`
+				}
+			};
+			// tslint:disable no-any
+			const keyframes: NotifierAnimationPresetKeyframes = slide.hide( <any> testNotification );
+			// tslint:enable no-any
+
+			expect( keyframes ).toEqual( expectedKeyframes );
+
+		} );
+
+	} );
+
+} );
 
 /**
  * Mock Notification Height
