+++ src/lib/src/components/notifier-container.component.spec.ts
@@ -15,44 +15,750 @@ import { NotifierService } from './../services/notifier.service';
 /**
  * Notifier Container Component - Unit Test
  */
-export function main(): void {
+describe( 'Notifier Container Component', () => {
 
-	describe( 'Notifier Container Component', () => {
+	let componentFixture: ComponentFixture<NotifierContainerComponent>;
+	let componentInstance: NotifierContainerComponent;
+	let queueService: MockNotifierQueueService;
 
-		let componentFixture: ComponentFixture<NotifierContainerComponent>;
-		let componentInstance: NotifierContainerComponent;
-		let queueService: MockNotifierQueueService;
+	it( 'should instantiate', () => {
 
-		it( 'should instantiate', () => {
+		// Setup test module
+		beforeEachWithConfig( new NotifierConfig() );
 
-			// Setup test module
-			beforeEachWithConfig( new NotifierConfig() );
+		expect( componentInstance ).toBeDefined();
 
-			expect( componentInstance ).toBeDefined();
+	} );
 
+	it( 'should render', () => {
+
+		// Setup test module
+		beforeEachWithConfig( new NotifierConfig() );
+		componentFixture.detectChanges();
+
+	} );
+
+	it( 'should ignore unknown actions', fakeAsync( () => {
+
+		// Setup test module
+		beforeEachWithConfig( new NotifierConfig() );
+		componentFixture.detectChanges();
+		jest.spyOn( queueService, 'continue' );
+
+		// tslint:disable no-any
+		queueService.push( <any> {
+			payload: 'STUFF',
+			type: 'WHATEVS'
 		} );
+		// tslint:enable no-any
+		componentFixture.detectChanges();
+		tick();
 
-		it( 'should render', () => {
+		expect( queueService.continue ).toHaveBeenCalled();
+
+	} ) );
+
+	describe( '(show)', () => {
+
+		it( 'should show the first notification', fakeAsync( () => {
 
 			// Setup test module
 			beforeEachWithConfig( new NotifierConfig() );
 			componentFixture.detectChanges();
+			jest.spyOn( queueService, 'continue' );
 
-		} );
-
-		it( 'should ignore unknown actions', fakeAsync( () => {
-
-			// Setup test module
-			beforeEachWithConfig( new NotifierConfig() );
-			componentFixture.detectChanges();
-			spyOn( queueService, 'continue' );
-
-			// tslint:disable no-any
-			queueService.push( <any> {
-				payload: 'STUFF',
-				type: 'WHATEVS'
+			queueService.push( {
+				payload: {
+					message: 'Lorem ipsum dolor sit amet.',
+					type: 'SUCCESS'
+				},
+				type: 'SHOW'
 			} );
+			componentFixture.detectChanges();
+			const listElements: Array<DebugElement> = componentFixture.debugElement.queryAll( By.css( '.x-notifier__container-list' ) );
+
+			expect( listElements.length ).toBe( 1 );
+
+			const mockNotificationComponent: MockNotifierNotificationComponent = new MockNotifierNotificationComponent();
+			jest.spyOn( mockNotificationComponent, 'show' ); // Continue
+			// tslint:disable no-any
+			componentInstance.onNotificationReady( <any> mockNotificationComponent ); // Trigger the ready event manually
 			// tslint:enable no-any
+
+			tick();
+
+			expect( mockNotificationComponent.show ).toHaveBeenCalled();
+			expect( queueService.continue ).toHaveBeenCalled();
+
+		} ) );
+
+		it( 'should switch out the old notification with the new one when stacking is disabled', fakeAsync( () => {
+
+			// Setup test module
+			beforeEachWithConfig( new NotifierConfig( {
+				behaviour: {
+					stacking: false
+				}
+			} ) );
+			componentFixture.detectChanges();
+
+			// Create the first notification
+			queueService.push( {
+				payload: {
+					message: 'Lorem ipsum dolor sit amet.',
+					type: 'SUCCESS'
+				},
+				type: 'SHOW'
+			} );
+			componentFixture.detectChanges();
+			const mockFirstNotificationComponent: MockNotifierNotificationComponent = new MockNotifierNotificationComponent();
+			jest.spyOn( mockFirstNotificationComponent, 'hide' ); // Continue
+			// tslint:disable no-any
+			componentInstance.onNotificationReady( <any> mockFirstNotificationComponent ); // Trigger the ready event manually
+			// tslint:enable no-any
+
+			tick();
+
+			// Create the second notification
+			queueService.push( {
+				payload: {
+					message: 'Blubberfish.',
+					type: 'ERROR'
+				},
+				type: 'SHOW'
+			} );
+			componentFixture.detectChanges();
+			const mockSecondNotificationComponent: MockNotifierNotificationComponent = new MockNotifierNotificationComponent();
+			jest.spyOn( mockSecondNotificationComponent, 'show' ); // Continue
+			// tslint:disable no-any
+			componentInstance.onNotificationReady( <any> mockSecondNotificationComponent ); // Trigger the ready event manually
+			// tslint:enable no-any
+
+			tick();
+
+			expect( mockFirstNotificationComponent.hide ).toHaveBeenCalled();
+			expect( mockSecondNotificationComponent.show ).toHaveBeenCalled();
+
+		} ) );
+
+		it( 'should hide and shift before showing the notification when stacking is enabled', fakeAsync( () => {
+
+			// Setup test module
+			beforeEachWithConfig( new NotifierConfig( {
+				animations: {
+					enabled: false
+				},
+				behaviour: {
+					stacking: 2
+				}
+			} ) );
+			componentFixture.detectChanges();
+
+			// Create the first notification
+			queueService.push( {
+				payload: {
+					message: 'Lorem ipsum dolor sit amet.',
+					type: 'SUCCESS'
+				},
+				type: 'SHOW'
+			} );
+			componentFixture.detectChanges();
+			const mockFirstNotificationComponent: MockNotifierNotificationComponent = new MockNotifierNotificationComponent();
+			jest.spyOn( mockFirstNotificationComponent, 'hide' ); // Continue
+			// tslint:disable no-any
+			componentInstance.onNotificationReady( <any> mockFirstNotificationComponent ); // Trigger the ready event manually
+			// tslint:enable no-any
+
+			tick();
+
+			// Create the second notification
+			queueService.push( {
+				payload: {
+					message: 'Blubberfish.',
+					type: 'ERROR'
+				},
+				type: 'SHOW'
+			} );
+			componentFixture.detectChanges();
+			const mockSecondNotificationComponent: MockNotifierNotificationComponent = new MockNotifierNotificationComponent();
+			jest.spyOn( mockSecondNotificationComponent, 'shift' ); // Continue
+			// tslint:disable no-any
+			componentInstance.onNotificationReady( <any> mockSecondNotificationComponent ); // Trigger the ready event manually
+			// tslint:enable no-any
+
+			tick();
+
+			// Create the third notification
+			queueService.push( {
+				payload: {
+					message: 'This. Is. Angular!',
+					type: 'INFO'
+				},
+				type: 'SHOW'
+			} );
+			componentFixture.detectChanges();
+			const mockThirdNotificationComponent: MockNotifierNotificationComponent = new MockNotifierNotificationComponent();
+			jest.spyOn( mockThirdNotificationComponent, 'show' ); // Continue
+			// tslint:disable no-any
+			componentInstance.onNotificationReady( <any> mockThirdNotificationComponent ); // Trigger the ready event manually
+			// tslint:enable no-any
+
+			tick();
+
+			expect( mockFirstNotificationComponent.hide ).toHaveBeenCalled();
+			expect( mockSecondNotificationComponent.shift ).toHaveBeenCalled();
+			expect( mockThirdNotificationComponent.show ).toHaveBeenCalled();
+
+		} ) );
+
+		it( 'should hide and shift before showing the notification, when stacking is enabled (with animations)', fakeAsync( () => {
+
+			// Setup test module
+			const testNotifierConfig: NotifierConfig = new NotifierConfig( {
+				animations: {
+					overlap: false
+				},
+				behaviour: {
+					stacking: 2
+				}
+			} );
+			beforeEachWithConfig( testNotifierConfig );
+			componentFixture.detectChanges();
+
+			// Create the first notification
+			queueService.push( {
+				payload: {
+					message: 'Lorem ipsum dolor sit amet.',
+					type: 'SUCCESS'
+				},
+				type: 'SHOW'
+			} );
+			componentFixture.detectChanges();
+			const mockFirstNotificationComponent: MockNotifierNotificationComponent = new MockNotifierNotificationComponent();
+			jest.spyOn( mockFirstNotificationComponent, 'hide' ); // Continue
+			// tslint:disable no-any
+			componentInstance.onNotificationReady( <any> mockFirstNotificationComponent ); // Trigger the ready event manually
+			// tslint:enable no-any
+
+			tick();
+
+			// Create the second notification
+			queueService.push( {
+				payload: {
+					message: 'Blubberfish.',
+					type: 'ERROR'
+				},
+				type: 'SHOW'
+			} );
+			componentFixture.detectChanges();
+			const mockSecondNotificationComponent: MockNotifierNotificationComponent = new MockNotifierNotificationComponent();
+			jest.spyOn( mockSecondNotificationComponent, 'shift' ); // Continue
+			// tslint:disable no-any
+			componentInstance.onNotificationReady( <any> mockSecondNotificationComponent ); // Trigger the ready event manually
+			// tslint:enable no-any
+
+			tick();
+
+			// Create the third notification
+			queueService.push( {
+				payload: {
+					message: 'This. Is. Angular!',
+					type: 'INFO'
+				},
+				type: 'SHOW'
+			} );
+			componentFixture.detectChanges();
+			const mockThirdNotificationComponent: MockNotifierNotificationComponent = new MockNotifierNotificationComponent();
+			jest.spyOn( mockThirdNotificationComponent, 'show' ); // Continue
+			// tslint:disable no-any
+			componentInstance.onNotificationReady( <any> mockThirdNotificationComponent ); // Trigger the ready event manually
+			// tslint:enable no-any
+
+			tick();
+
+			expect( mockFirstNotificationComponent.hide ).toHaveBeenCalled();
+			expect( mockSecondNotificationComponent.shift ).toHaveBeenCalled();
+			expect( mockThirdNotificationComponent.show ).toHaveBeenCalled();
+
+		} ) );
+
+		it( 'should hide and shift before showing the notification, when tacking is enabled (with overlapping animations)', fakeAsync( () => {
+
+			// Setup test module
+			const testNotifierConfig: NotifierConfig = new NotifierConfig( {
+				behaviour: {
+					stacking: 2
+				}
+			} );
+			beforeEachWithConfig( testNotifierConfig );
+			componentFixture.detectChanges();
+
+			// Create the first notification
+			queueService.push( {
+				payload: {
+					message: 'Lorem ipsum dolor sit amet.',
+					type: 'SUCCESS'
+				},
+				type: 'SHOW'
+			} );
+			componentFixture.detectChanges();
+			const mockFirstNotificationComponent: MockNotifierNotificationComponent = new MockNotifierNotificationComponent();
+			jest.spyOn( mockFirstNotificationComponent, 'hide' ); // Continue
+			// tslint:disable no-any
+			componentInstance.onNotificationReady( <any> mockFirstNotificationComponent ); // Trigger the ready event manually
+			// tslint:enable no-any
+
+			tick();
+
+			// Create the second notification
+			queueService.push( {
+				payload: {
+					message: 'Blubberfish.',
+					type: 'ERROR'
+				},
+				type: 'SHOW'
+			} );
+			componentFixture.detectChanges();
+			const mockSecondNotificationComponent: MockNotifierNotificationComponent = new MockNotifierNotificationComponent();
+			jest.spyOn( mockSecondNotificationComponent, 'shift' ); // Continue
+			// tslint:disable no-any
+			componentInstance.onNotificationReady( <any> mockSecondNotificationComponent ); // Trigger the ready event manually
+			// tslint:enable no-any
+
+			tick();
+
+			// Create the third notification
+			queueService.push( {
+				payload: {
+					message: 'This. Is. Angular!',
+					type: 'INFO'
+				},
+				type: 'SHOW'
+			} );
+			componentFixture.detectChanges();
+			const mockThirdNotificationComponent: MockNotifierNotificationComponent = new MockNotifierNotificationComponent();
+			jest.spyOn( mockThirdNotificationComponent, 'show' ); // Continue
+			// tslint:disable no-any
+			componentInstance.onNotificationReady( <any> mockThirdNotificationComponent ); // Trigger the ready event manually
+			// tslint:enable no-any
+
+			tick( testNotifierConfig.animations.hide.speed + testNotifierConfig.animations.shift.speed
+				- <number> testNotifierConfig.animations.overlap );
+
+			expect( mockFirstNotificationComponent.hide ).toHaveBeenCalled();
+			expect( mockSecondNotificationComponent.shift ).toHaveBeenCalled();
+			expect( mockThirdNotificationComponent.show ).toHaveBeenCalled();
+
+		} ) );
+
+	} );
+
+	describe( '(hide)', () => {
+
+		it( 'should hide one notification', fakeAsync( () => {
+
+			// Setup test module
+			beforeEachWithConfig( new NotifierConfig( {
+				animations: {
+					enabled: false
+				}
+			} ) );
+			componentFixture.detectChanges();
+			jest.spyOn( queueService, 'continue' );
+
+			const testNotificationId: string = 'FAKE_ID';
+
+			// Show notification
+			queueService.push( {
+				payload: {
+					id: testNotificationId,
+					message: 'Lorem ipsum dolor sit amet.',
+					type: 'SUCCESS'
+				},
+				type: 'SHOW'
+			} );
+			componentFixture.detectChanges();
+			const mockNotificationComponent: MockNotifierNotificationComponent = new MockNotifierNotificationComponent();
+			jest.spyOn( mockNotificationComponent, 'hide' ); // Continue
+			// tslint:disable no-any
+			componentInstance.onNotificationReady( <any> mockNotificationComponent ); // Trigger the ready event manually
+			// tslint:enable no-any
+
+			tick();
+
+			// Hide notification
+			queueService.push( {
+				payload: testNotificationId,
+				type: 'HIDE'
+			} );
+			componentFixture.detectChanges();
+
+			tick();
+			componentFixture.detectChanges(); // Run a second change detection (to update the template)
+
+			const listElements: Array<DebugElement> = componentFixture.debugElement.queryAll( By.css( '.x-notifier__container-list' ) );
+
+			expect( listElements.length ).toBe( 0 );
+			expect( mockNotificationComponent.hide ).toHaveBeenCalled();
+			expect( queueService.continue ).toHaveBeenCalled();
+
+		} ) );
+
+		it( 'should skip if the notification to hide does not exist', fakeAsync( () => {
+
+			// Setup test module
+			beforeEachWithConfig( new NotifierConfig( {
+				animations: {
+					enabled: false
+				}
+			} ) );
+			componentFixture.detectChanges();
+			jest.spyOn( queueService, 'continue' );
+
+			// Show first notification
+			queueService.push( {
+				payload: {
+					id: 'FAKE_ID',
+					message: 'Lorem ipsum dolor sit amet.',
+					type: 'SUCCESS'
+				},
+				type: 'SHOW'
+			} );
+			componentFixture.detectChanges();
+			const mockNotificationComponent: MockNotifierNotificationComponent = new MockNotifierNotificationComponent();
+			jest.spyOn( mockNotificationComponent, 'hide' ); // Continue
+			// tslint:disable no-any
+			componentInstance.onNotificationReady( <any> mockNotificationComponent ); // Trigger the ready event manually
+			// tslint:enable no-any
+
+			tick();
+
+			// Hide notification
+			queueService.push( {
+				payload: 'NOT_EXISTING_ID',
+				type: 'HIDE'
+			} );
+			componentFixture.detectChanges();
+			tick();
+			componentFixture.detectChanges(); // Run a second change detection (to update the template)
+
+			let listElements: Array<DebugElement> = componentFixture.debugElement.queryAll( By.css( '.x-notifier__container-list' ) );
+
+			expect( listElements.length ).toBe( 1 );
+			expect( mockNotificationComponent.hide ).not.toHaveBeenCalled();
+			expect( queueService.continue ).toHaveBeenCalled();
+
+		} ) );
+
+		it( 'should shift before hiding the notification if necessary', fakeAsync( () => {
+
+			// Setup test module
+			beforeEachWithConfig( new NotifierConfig( {
+				animations: {
+					enabled: false
+				}
+			} ) );
+			componentFixture.detectChanges();
+			jest.spyOn( queueService, 'continue' );
+
+			// Show first notification
+			queueService.push( {
+				payload: {
+					message: 'Whut.',
+					type: 'ERROR'
+				},
+				type: 'SHOW'
+			} );
+			componentFixture.detectChanges();
+			const mockFirstNotificationComponent: MockNotifierNotificationComponent = new MockNotifierNotificationComponent();
+			jest.spyOn( mockFirstNotificationComponent, 'shift' ); // Continue
+			// tslint:disable no-any
+			componentInstance.onNotificationReady( <any> mockFirstNotificationComponent ); // Trigger the ready event manually
+			// tslint:enable no-any
+
+			// Show second notification
+			const testNotificationId: string = 'FAKE_ID';
+			queueService.push( {
+				payload: {
+					id: testNotificationId,
+					message: 'Lorem ipsum dolor sit amet.',
+					type: 'SUCCESS'
+				},
+				type: 'SHOW'
+			} );
+			componentFixture.detectChanges();
+			const mockSecondNotificationComponent: MockNotifierNotificationComponent = new MockNotifierNotificationComponent();
+			jest.spyOn( mockSecondNotificationComponent, 'hide' ); // Continue
+			// tslint:disable no-any
+			componentInstance.onNotificationReady( <any> mockSecondNotificationComponent ); // Trigger the ready event manually
+			// tslint:enable no-any
+
+			// Hide second notification
+			queueService.push( {
+				payload: testNotificationId,
+				type: 'HIDE'
+			} );
+			componentFixture.detectChanges();
+
+			tick();
+			componentFixture.detectChanges(); // Run a second change detection (to update the template)
+
+			const listElements: Array<DebugElement> = componentFixture.debugElement.queryAll( By.css( '.x-notifier__container-list' ) );
+
+			const expectedCallTimes: number = 3;
+			expect( listElements.length ).toBe( 1 );
+			expect( mockFirstNotificationComponent.shift ).toHaveBeenCalled();
+			expect( mockSecondNotificationComponent.hide ).toHaveBeenCalled();
+			expect( queueService.continue ).toHaveBeenCalledTimes( expectedCallTimes );
+
+		} ) );
+
+		it( 'should shift before hiding the notification if necessary (with animations)', fakeAsync( () => {
+
+			// Setup test module
+			const testNotifierConfig: NotifierConfig = new NotifierConfig( {
+				animations: {
+					overlap: false
+				}
+			} );
+			beforeEachWithConfig( testNotifierConfig );
+			componentFixture.detectChanges();
+			jest.spyOn( queueService, 'continue' );
+
+			// Show first notification
+			queueService.push( {
+				payload: {
+					message: 'Whut.',
+					type: 'ERROR'
+				},
+				type: 'SHOW'
+			} );
+			componentFixture.detectChanges();
+			const mockFirstNotificationComponent: MockNotifierNotificationComponent = new MockNotifierNotificationComponent();
+			jest.spyOn( mockFirstNotificationComponent, 'shift' ); // Continue
+			// tslint:disable no-any
+			componentInstance.onNotificationReady( <any> mockFirstNotificationComponent ); // Trigger the ready event manually
+			// tslint:enable no-any
+
+			// Show second notification
+			const testNotificationId: string = 'FAKE_ID';
+			queueService.push( {
+				payload: {
+					id: testNotificationId,
+					message: 'Lorem ipsum dolor sit amet.',
+					type: 'SUCCESS'
+				},
+				type: 'SHOW'
+			} );
+			componentFixture.detectChanges();
+			const mockSecondNotificationComponent: MockNotifierNotificationComponent = new MockNotifierNotificationComponent();
+			jest.spyOn( mockSecondNotificationComponent, 'hide' ); // Continue
+			// tslint:disable no-any
+			componentInstance.onNotificationReady( <any> mockSecondNotificationComponent ); // Trigger the ready event manually
+			// tslint:enable no-any
+
+			// Hide second notification
+			queueService.push( {
+				payload: testNotificationId,
+				type: 'HIDE'
+			} );
+			componentFixture.detectChanges();
+
+			tick( testNotifierConfig.animations.hide.speed );
+			componentFixture.detectChanges(); // Run a second change detection (to update the template)
+
+			const listElements: Array<DebugElement> = componentFixture.debugElement.queryAll( By.css( '.x-notifier__container-list' ) );
+
+			const expectedCallTimes: number = 3;
+			expect( listElements.length ).toBe( 1 );
+			expect( mockFirstNotificationComponent.shift ).toHaveBeenCalled();
+			expect( mockSecondNotificationComponent.hide ).toHaveBeenCalled();
+			expect( queueService.continue ).toHaveBeenCalledTimes( expectedCallTimes );
+
+		} ) );
+
+		it( 'should shift before hiding the notification if necessary (with overlapping animations)', fakeAsync( () => {
+
+			// Setup test module
+			const testNotifierConfig: NotifierConfig = new NotifierConfig();
+			beforeEachWithConfig( testNotifierConfig );
+			componentFixture.detectChanges();
+			jest.spyOn( queueService, 'continue' );
+
+			// Show first notification
+			queueService.push( {
+				payload: {
+					message: 'Whut.',
+					type: 'ERROR'
+				},
+				type: 'SHOW'
+			} );
+			componentFixture.detectChanges();
+			const mockFirstNotificationComponent: MockNotifierNotificationComponent = new MockNotifierNotificationComponent();
+			jest.spyOn( mockFirstNotificationComponent, 'shift' ); // Continue
+			// tslint:disable no-any
+			componentInstance.onNotificationReady( <any> mockFirstNotificationComponent ); // Trigger the ready event manually
+			// tslint:enable no-any
+
+			// Show second notification
+			const testNotificationId: string = 'FAKE_ID';
+			queueService.push( {
+				payload: {
+					id: testNotificationId,
+					message: 'Lorem ipsum dolor sit amet.',
+					type: 'SUCCESS'
+				},
+				type: 'SHOW'
+			} );
+			componentFixture.detectChanges();
+			const mockSecondNotificationComponent: MockNotifierNotificationComponent = new MockNotifierNotificationComponent();
+			jest.spyOn( mockSecondNotificationComponent, 'hide' ); // Continue
+			// tslint:disable no-any
+			componentInstance.onNotificationReady( <any> mockSecondNotificationComponent ); // Trigger the ready event manually
+			// tslint:enable no-any
+
+			// Hide second notification
+			queueService.push( {
+				payload: testNotificationId,
+				type: 'HIDE'
+			} );
+			componentFixture.detectChanges();
+
+			tick( testNotifierConfig.animations.hide.speed - <number> testNotifierConfig.animations.overlap );
+			componentFixture.detectChanges(); // Run a second change detection (to update the template)
+
+			const listElements: Array<DebugElement> = componentFixture.debugElement.queryAll( By.css( '.x-notifier__container-list' ) );
+
+			const expectedCallTimes: number = 3;
+			expect( listElements.length ).toBe( 1 );
+			expect( mockFirstNotificationComponent.shift ).toHaveBeenCalled();
+			expect( mockSecondNotificationComponent.hide ).toHaveBeenCalled();
+			expect( queueService.continue ).toHaveBeenCalledTimes( expectedCallTimes );
+
+		} ) );
+
+		it( 'should hide the notification when the dismiss event gets dispatched', fakeAsync( () => {
+
+			// Setup test module
+			beforeEachWithConfig( new NotifierConfig( {
+				animations: {
+					enabled: false
+				}
+			} ) );
+			componentFixture.detectChanges();
+			jest.spyOn( queueService, 'continue' );
+
+			// Show second notification
+			const testNotificationId: string = 'FAKE_ID';
+			queueService.push( {
+				payload: {
+					id: testNotificationId,
+					message: 'Lorem ipsum dolor sit amet.',
+					type: 'SUCCESS'
+				},
+				type: 'SHOW'
+			} );
+			componentFixture.detectChanges();
+			const mockNotificationComponent: MockNotifierNotificationComponent = new MockNotifierNotificationComponent();
+			jest.spyOn( mockNotificationComponent, 'hide' ); // Continue
+			// tslint:disable no-any
+			componentInstance.onNotificationReady( <any> mockNotificationComponent ); // Trigger the ready event manually
+			// tslint:enable no-any
+
+			// Hide second notification
+			componentInstance.onNotificationDismiss( testNotificationId );
+			tick();
+			componentFixture.detectChanges();
+
+			const listElements: Array<DebugElement> = componentFixture.debugElement.queryAll( By.css( '.x-notifier__container-list' ) );
+
+			const expectedCallTimes: number = 2;
+			expect( listElements.length ).toBe( 0 );
+			expect( mockNotificationComponent.hide ).toHaveBeenCalled();
+			expect( queueService.continue ).toHaveBeenCalledTimes( expectedCallTimes );
+
+		} ) );
+
+	} );
+
+	describe( '(hide oldest / newest)', () => {
+
+		it( 'should hide the oldest notification', fakeAsync( () => {
+
+			// Setup test module
+			beforeEachWithConfig( new NotifierConfig( {
+				animations: {
+					enabled: false
+				}
+			} ) );
+			componentFixture.detectChanges();
+			jest.spyOn( queueService, 'continue' );
+
+			// Show first notification
+			queueService.push( {
+				payload: {
+					id: 'FAKE_ID',
+					message: 'Lorem ipsum dolor sit amet.',
+					type: 'SUCCESS'
+				},
+				type: 'SHOW'
+			} );
+			componentFixture.detectChanges();
+			const mockFirstNotificationComponent: MockNotifierNotificationComponent = new MockNotifierNotificationComponent();
+			jest.spyOn( mockFirstNotificationComponent, 'hide' ); // Continue
+			// tslint:disable no-any
+			componentInstance.onNotificationReady( <any> mockFirstNotificationComponent ); // Trigger the ready event manually
+			// tslint:enable no-any
+
+			// Show first notification
+			queueService.push( {
+				payload: {
+					message: 'Whut.',
+					type: 'ERROR'
+				},
+				type: 'SHOW'
+			} );
+			componentFixture.detectChanges();
+			const mockSecondNotificationComponent: MockNotifierNotificationComponent = new MockNotifierNotificationComponent();
+			jest.spyOn( mockSecondNotificationComponent, 'hide' ); // Continue
+			// tslint:disable no-any
+			componentInstance.onNotificationReady( <any> mockSecondNotificationComponent ); // Trigger the ready event manually
+			// tslint:enable no-any
+
+			// Hide second notification
+			queueService.push( {
+				type: 'HIDE_OLDEST'
+			} );
+			componentFixture.detectChanges();
+
+			tick();
+			componentFixture.detectChanges(); // Run a second change detection (to update the template)
+
+			const listElements: Array<DebugElement> = componentFixture.debugElement.queryAll( By.css( '.x-notifier__container-list' ) );
+
+			const expectedCallTimes: number = 3;
+			expect( listElements.length ).toBe( 1 );
+			expect( mockFirstNotificationComponent.hide ).toHaveBeenCalled();
+			expect( mockSecondNotificationComponent.hide ).not.toHaveBeenCalled();
+			expect( queueService.continue ).toHaveBeenCalledTimes( expectedCallTimes );
+
+		} ) );
+
+		it( 'should skip hiding the oldest notification if there are no notifications', fakeAsync( () => {
+
+			// Setup test module
+			beforeEachWithConfig( new NotifierConfig( {
+				animations: {
+					enabled: false
+				}
+			} ) );
+			componentFixture.detectChanges();
+			jest.spyOn( queueService, 'continue' );
+
+			// Hide notification
+			queueService.push( {
+				type: 'HIDE_OLDEST'
+			} );
 			componentFixture.detectChanges();
 			tick();
 
@@ -60,978 +766,268 @@ export function main(): void {
 
 		} ) );
 
-		describe( '(show)', () => {
+		it( 'should hide the newest notification', fakeAsync( () => {
 
-			it( 'should show the first notification', fakeAsync( () => {
-
-				// Setup test module
-				beforeEachWithConfig( new NotifierConfig() );
-				componentFixture.detectChanges();
-				spyOn( queueService, 'continue' );
-
-				queueService.push( {
-					payload: {
-						message: 'Lorem ipsum dolor sit amet.',
-						type: 'SUCCESS'
-					},
-					type: 'SHOW'
-				} );
-				componentFixture.detectChanges();
-				const listElements: Array<DebugElement> = componentFixture.debugElement.queryAll( By.css( '.x-notifier__container-list' ) );
-
-				expect( listElements.length ).toBe( 1 );
-
-				const mockNotificationComponent: MockNotifierNotificationComponent = new MockNotifierNotificationComponent();
-				spyOn( mockNotificationComponent, 'show' ).and.callThrough(); // Continue
-				// tslint:disable no-any
-				componentInstance.onNotificationReady( <any> mockNotificationComponent ); // Trigger the ready event manually
-				// tslint:enable no-any
-
-				tick();
-
-				expect( mockNotificationComponent.show ).toHaveBeenCalled();
-				expect( queueService.continue ).toHaveBeenCalled();
-
+			// Setup test module
+			beforeEachWithConfig( new NotifierConfig( {
+				animations: {
+					enabled: false
+				}
 			} ) );
-
-			it( 'should switch out the old notification with the new one when stacking is disabled', fakeAsync( () => {
-
-				// Setup test module
-				beforeEachWithConfig( new NotifierConfig( {
-					behaviour: {
-						stacking: false
-					}
-				} ) );
-				componentFixture.detectChanges();
-
-				// Create the first notification
-				queueService.push( {
-					payload: {
-						message: 'Lorem ipsum dolor sit amet.',
-						type: 'SUCCESS'
-					},
-					type: 'SHOW'
-				} );
-				componentFixture.detectChanges();
-				const mockFirstNotificationComponent: MockNotifierNotificationComponent = new MockNotifierNotificationComponent();
-				spyOn( mockFirstNotificationComponent, 'hide' ).and.callThrough(); // Continue
-				// tslint:disable no-any
-				componentInstance.onNotificationReady( <any> mockFirstNotificationComponent ); // Trigger the ready event manually
-				// tslint:enable no-any
-
-				tick();
-
-				// Create the second notification
-				queueService.push( {
-					payload: {
-						message: 'Blubberfish.',
-						type: 'ERROR'
-					},
-					type: 'SHOW'
-				} );
-				componentFixture.detectChanges();
-				const mockSecondNotificationComponent: MockNotifierNotificationComponent = new MockNotifierNotificationComponent();
-				spyOn( mockSecondNotificationComponent, 'show' ).and.callThrough(); // Continue
-				// tslint:disable no-any
-				componentInstance.onNotificationReady( <any> mockSecondNotificationComponent ); // Trigger the ready event manually
-				// tslint:enable no-any
-
-				tick();
-
-				expect( mockFirstNotificationComponent.hide ).toHaveBeenCalled();
-				expect( mockSecondNotificationComponent.show ).toHaveBeenCalled();
-
-			} ) );
-
-			it( 'should hide and shift before showing the notification when stacking is enabled', fakeAsync( () => {
-
-				// Setup test module
-				beforeEachWithConfig( new NotifierConfig( {
-					animations: {
-						enabled: false
-					},
-					behaviour: {
-						stacking: 2
-					}
-				} ) );
-				componentFixture.detectChanges();
-
-				// Create the first notification
-				queueService.push( {
-					payload: {
-						message: 'Lorem ipsum dolor sit amet.',
-						type: 'SUCCESS'
-					},
-					type: 'SHOW'
-				} );
-				componentFixture.detectChanges();
-				const mockFirstNotificationComponent: MockNotifierNotificationComponent = new MockNotifierNotificationComponent();
-				spyOn( mockFirstNotificationComponent, 'hide' ).and.callThrough(); // Continue
-				// tslint:disable no-any
-				componentInstance.onNotificationReady( <any> mockFirstNotificationComponent ); // Trigger the ready event manually
-				// tslint:enable no-any
-
-				tick();
-
-				// Create the second notification
-				queueService.push( {
-					payload: {
-						message: 'Blubberfish.',
-						type: 'ERROR'
-					},
-					type: 'SHOW'
-				} );
-				componentFixture.detectChanges();
-				const mockSecondNotificationComponent: MockNotifierNotificationComponent = new MockNotifierNotificationComponent();
-				spyOn( mockSecondNotificationComponent, 'shift' ).and.callThrough(); // Continue
-				// tslint:disable no-any
-				componentInstance.onNotificationReady( <any> mockSecondNotificationComponent ); // Trigger the ready event manually
-				// tslint:enable no-any
-
-				tick();
-
-				// Create the third notification
-				queueService.push( {
-					payload: {
-						message: 'This. Is. Angular!',
-						type: 'INFO'
-					},
-					type: 'SHOW'
-				} );
-				componentFixture.detectChanges();
-				const mockThirdNotificationComponent: MockNotifierNotificationComponent = new MockNotifierNotificationComponent();
-				spyOn( mockThirdNotificationComponent, 'show' ).and.callThrough(); // Continue
-				// tslint:disable no-any
-				componentInstance.onNotificationReady( <any> mockThirdNotificationComponent ); // Trigger the ready event manually
-				// tslint:enable no-any
-
-				tick();
-
-				expect( mockFirstNotificationComponent.hide ).toHaveBeenCalled();
-				expect( mockSecondNotificationComponent.shift ).toHaveBeenCalled();
-				expect( mockThirdNotificationComponent.show ).toHaveBeenCalled();
-
-			} ) );
-
-			it( 'should hide and shift before showing the notification, when stacking is enabled (with animations)', fakeAsync( () => {
-
-				// Setup test module
-				const testNotifierConfig: NotifierConfig = new NotifierConfig( {
-					animations: {
-						overlap: false
-					},
-					behaviour: {
-						stacking: 2
-					}
-				} );
-				beforeEachWithConfig( testNotifierConfig );
-				componentFixture.detectChanges();
-
-				// Create the first notification
-				queueService.push( {
-					payload: {
-						message: 'Lorem ipsum dolor sit amet.',
-						type: 'SUCCESS'
-					},
-					type: 'SHOW'
-				} );
-				componentFixture.detectChanges();
-				const mockFirstNotificationComponent: MockNotifierNotificationComponent = new MockNotifierNotificationComponent();
-				spyOn( mockFirstNotificationComponent, 'hide' ).and.callThrough(); // Continue
-				// tslint:disable no-any
-				componentInstance.onNotificationReady( <any> mockFirstNotificationComponent ); // Trigger the ready event manually
-				// tslint:enable no-any
-
-				tick();
-
-				// Create the second notification
-				queueService.push( {
-					payload: {
-						message: 'Blubberfish.',
-						type: 'ERROR'
-					},
-					type: 'SHOW'
-				} );
-				componentFixture.detectChanges();
-				const mockSecondNotificationComponent: MockNotifierNotificationComponent = new MockNotifierNotificationComponent();
-				spyOn( mockSecondNotificationComponent, 'shift' ).and.callThrough(); // Continue
-				// tslint:disable no-any
-				componentInstance.onNotificationReady( <any> mockSecondNotificationComponent ); // Trigger the ready event manually
-				// tslint:enable no-any
-
-				tick();
-
-				// Create the third notification
-				queueService.push( {
-					payload: {
-						message: 'This. Is. Angular!',
-						type: 'INFO'
-					},
-					type: 'SHOW'
-				} );
-				componentFixture.detectChanges();
-				const mockThirdNotificationComponent: MockNotifierNotificationComponent = new MockNotifierNotificationComponent();
-				spyOn( mockThirdNotificationComponent, 'show' ).and.callThrough(); // Continue
-				// tslint:disable no-any
-				componentInstance.onNotificationReady( <any> mockThirdNotificationComponent ); // Trigger the ready event manually
-				// tslint:enable no-any
-
-				tick();
-
-				expect( mockFirstNotificationComponent.hide ).toHaveBeenCalled();
-				expect( mockSecondNotificationComponent.shift ).toHaveBeenCalled();
-				expect( mockThirdNotificationComponent.show ).toHaveBeenCalled();
-
-			} ) );
-
-			it( 'should hide and shift before showing the notification, when tacking is enabled (with overlapping animations)', fakeAsync( () => {
-
-				// Setup test module
-				const testNotifierConfig: NotifierConfig = new NotifierConfig( {
-					behaviour: {
-						stacking: 2
-					}
-				} );
-				beforeEachWithConfig( testNotifierConfig );
-				componentFixture.detectChanges();
-
-				// Create the first notification
-				queueService.push( {
-					payload: {
-						message: 'Lorem ipsum dolor sit amet.',
-						type: 'SUCCESS'
-					},
-					type: 'SHOW'
-				} );
-				componentFixture.detectChanges();
-				const mockFirstNotificationComponent: MockNotifierNotificationComponent = new MockNotifierNotificationComponent();
-				spyOn( mockFirstNotificationComponent, 'hide' ).and.callThrough(); // Continue
-				// tslint:disable no-any
-				componentInstance.onNotificationReady( <any> mockFirstNotificationComponent ); // Trigger the ready event manually
-				// tslint:enable no-any
-
-				tick();
-
-				// Create the second notification
-				queueService.push( {
-					payload: {
-						message: 'Blubberfish.',
-						type: 'ERROR'
-					},
-					type: 'SHOW'
-				} );
-				componentFixture.detectChanges();
-				const mockSecondNotificationComponent: MockNotifierNotificationComponent = new MockNotifierNotificationComponent();
-				spyOn( mockSecondNotificationComponent, 'shift' ).and.callThrough(); // Continue
-				// tslint:disable no-any
-				componentInstance.onNotificationReady( <any> mockSecondNotificationComponent ); // Trigger the ready event manually
-				// tslint:enable no-any
-
-				tick();
-
-				// Create the third notification
-				queueService.push( {
-					payload: {
-						message: 'This. Is. Angular!',
-						type: 'INFO'
-					},
-					type: 'SHOW'
-				} );
-				componentFixture.detectChanges();
-				const mockThirdNotificationComponent: MockNotifierNotificationComponent = new MockNotifierNotificationComponent();
-				spyOn( mockThirdNotificationComponent, 'show' ).and.callThrough(); // Continue
-				// tslint:disable no-any
-				componentInstance.onNotificationReady( <any> mockThirdNotificationComponent ); // Trigger the ready event manually
-				// tslint:enable no-any
-
-				tick( testNotifierConfig.animations.hide.speed + testNotifierConfig.animations.shift.speed
-					- <number> testNotifierConfig.animations.overlap );
-
-				expect( mockFirstNotificationComponent.hide ).toHaveBeenCalled();
-				expect( mockSecondNotificationComponent.shift ).toHaveBeenCalled();
-				expect( mockThirdNotificationComponent.show ).toHaveBeenCalled();
-
-			} ) );
-
-		} );
-
-		describe( '(hide)', () => {
-
-			it( 'should hide one notification', fakeAsync( () => {
-
-				// Setup test module
-				beforeEachWithConfig( new NotifierConfig( {
-					animations: {
-						enabled: false
-					}
-				} ) );
-				componentFixture.detectChanges();
-				spyOn( queueService, 'continue' );
-
-				const testNotificationId: string = 'FAKE_ID';
-
-				// Show notification
-				queueService.push( {
-					payload: {
-						id: testNotificationId,
-						message: 'Lorem ipsum dolor sit amet.',
-						type: 'SUCCESS'
-					},
-					type: 'SHOW'
-				} );
-				componentFixture.detectChanges();
-				const mockNotificationComponent: MockNotifierNotificationComponent = new MockNotifierNotificationComponent();
-				spyOn( mockNotificationComponent, 'hide' ).and.callThrough(); // Continue
-				// tslint:disable no-any
-				componentInstance.onNotificationReady( <any> mockNotificationComponent ); // Trigger the ready event manually
-				// tslint:enable no-any
-
-				tick();
-
-				// Hide notification
-				queueService.push( {
-					payload: testNotificationId,
-					type: 'HIDE'
-				} );
-				componentFixture.detectChanges();
-
-				tick();
-				componentFixture.detectChanges(); // Run a second change detection (to update the template)
-
-				const listElements: Array<DebugElement> = componentFixture.debugElement.queryAll( By.css( '.x-notifier__container-list' ) );
-
-				expect( listElements.length ).toBe( 0 );
-				expect( mockNotificationComponent.hide ).toHaveBeenCalled();
-				expect( queueService.continue ).toHaveBeenCalled();
-
-			} ) );
-
-			it( 'should skip if the notification to hide does not exist', fakeAsync( () => {
-
-				// Setup test module
-				beforeEachWithConfig( new NotifierConfig( {
-					animations: {
-						enabled: false
-					}
-				} ) );
-				componentFixture.detectChanges();
-				spyOn( queueService, 'continue' );
-
-				// Show first notification
-				queueService.push( {
-					payload: {
-						id: 'FAKE_ID',
-						message: 'Lorem ipsum dolor sit amet.',
-						type: 'SUCCESS'
-					},
-					type: 'SHOW'
-				} );
-				componentFixture.detectChanges();
-				const mockNotificationComponent: MockNotifierNotificationComponent = new MockNotifierNotificationComponent();
-				spyOn( mockNotificationComponent, 'hide' ).and.callThrough(); // Continue
-				// tslint:disable no-any
-				componentInstance.onNotificationReady( <any> mockNotificationComponent ); // Trigger the ready event manually
-				// tslint:enable no-any
-
-				tick();
-
-				// Hide notification
-				queueService.push( {
-					payload: 'NOT_EXISTING_ID',
-					type: 'HIDE'
-				} );
-				componentFixture.detectChanges();
-				tick();
-				componentFixture.detectChanges(); // Run a second change detection (to update the template)
-
-				let listElements: Array<DebugElement> = componentFixture.debugElement.queryAll( By.css( '.x-notifier__container-list' ) );
-
-				expect( listElements.length ).toBe( 1 );
-				expect( mockNotificationComponent.hide ).not.toHaveBeenCalled();
-				expect( queueService.continue ).toHaveBeenCalled();
-
-			} ) );
-
-			it( 'should shift before hiding the notification if necessary', fakeAsync( () => {
-
-				// Setup test module
-				beforeEachWithConfig( new NotifierConfig( {
-					animations: {
-						enabled: false
-					}
-				} ) );
-				componentFixture.detectChanges();
-				spyOn( queueService, 'continue' );
-
-				// Show first notification
-				queueService.push( {
-					payload: {
-						message: 'Whut.',
-						type: 'ERROR'
-					},
-					type: 'SHOW'
-				} );
-				componentFixture.detectChanges();
-				const mockFirstNotificationComponent: MockNotifierNotificationComponent = new MockNotifierNotificationComponent();
-				spyOn( mockFirstNotificationComponent, 'shift' ).and.callThrough(); // Continue
-				// tslint:disable no-any
-				componentInstance.onNotificationReady( <any> mockFirstNotificationComponent ); // Trigger the ready event manually
-				// tslint:enable no-any
-
-				// Show second notification
-				const testNotificationId: string = 'FAKE_ID';
-				queueService.push( {
-					payload: {
-						id: testNotificationId,
-						message: 'Lorem ipsum dolor sit amet.',
-						type: 'SUCCESS'
-					},
-					type: 'SHOW'
-				} );
-				componentFixture.detectChanges();
-				const mockSecondNotificationComponent: MockNotifierNotificationComponent = new MockNotifierNotificationComponent();
-				spyOn( mockSecondNotificationComponent, 'hide' ).and.callThrough(); // Continue
-				// tslint:disable no-any
-				componentInstance.onNotificationReady( <any> mockSecondNotificationComponent ); // Trigger the ready event manually
-				// tslint:enable no-any
-
-				// Hide second notification
-				queueService.push( {
-					payload: testNotificationId,
-					type: 'HIDE'
-				} );
-				componentFixture.detectChanges();
-
-				tick();
-				componentFixture.detectChanges(); // Run a second change detection (to update the template)
-
-				const listElements: Array<DebugElement> = componentFixture.debugElement.queryAll( By.css( '.x-notifier__container-list' ) );
-
-				const expectedCallTimes: number = 3;
-				expect( listElements.length ).toBe( 1 );
-				expect( mockFirstNotificationComponent.shift ).toHaveBeenCalled();
-				expect( mockSecondNotificationComponent.hide ).toHaveBeenCalled();
-				expect( queueService.continue ).toHaveBeenCalledTimes( expectedCallTimes );
-
-			} ) );
-
-			it( 'should shift before hiding the notification if necessary (with animations)', fakeAsync( () => {
-
-				// Setup test module
-				const testNotifierConfig: NotifierConfig = new NotifierConfig( {
-					animations: {
-						overlap: false
-					}
-				} );
-				beforeEachWithConfig( testNotifierConfig );
-				componentFixture.detectChanges();
-				spyOn( queueService, 'continue' );
-
-				// Show first notification
-				queueService.push( {
-					payload: {
-						message: 'Whut.',
-						type: 'ERROR'
-					},
-					type: 'SHOW'
-				} );
-				componentFixture.detectChanges();
-				const mockFirstNotificationComponent: MockNotifierNotificationComponent = new MockNotifierNotificationComponent();
-				spyOn( mockFirstNotificationComponent, 'shift' ).and.callThrough(); // Continue
-				// tslint:disable no-any
-				componentInstance.onNotificationReady( <any> mockFirstNotificationComponent ); // Trigger the ready event manually
-				// tslint:enable no-any
-
-				// Show second notification
-				const testNotificationId: string = 'FAKE_ID';
-				queueService.push( {
-					payload: {
-						id: testNotificationId,
-						message: 'Lorem ipsum dolor sit amet.',
-						type: 'SUCCESS'
-					},
-					type: 'SHOW'
-				} );
-				componentFixture.detectChanges();
-				const mockSecondNotificationComponent: MockNotifierNotificationComponent = new MockNotifierNotificationComponent();
-				spyOn( mockSecondNotificationComponent, 'hide' ).and.callThrough(); // Continue
-				// tslint:disable no-any
-				componentInstance.onNotificationReady( <any> mockSecondNotificationComponent ); // Trigger the ready event manually
-				// tslint:enable no-any
-
-				// Hide second notification
-				queueService.push( {
-					payload: testNotificationId,
-					type: 'HIDE'
-				} );
-				componentFixture.detectChanges();
-
-				tick( testNotifierConfig.animations.hide.speed );
-				componentFixture.detectChanges(); // Run a second change detection (to update the template)
-
-				const listElements: Array<DebugElement> = componentFixture.debugElement.queryAll( By.css( '.x-notifier__container-list' ) );
-
-				const expectedCallTimes: number = 3;
-				expect( listElements.length ).toBe( 1 );
-				expect( mockFirstNotificationComponent.shift ).toHaveBeenCalled();
-				expect( mockSecondNotificationComponent.hide ).toHaveBeenCalled();
-				expect( queueService.continue ).toHaveBeenCalledTimes( expectedCallTimes );
-
-			} ) );
-
-			it( 'should shift before hiding the notification if necessary (with overlapping animations)', fakeAsync( () => {
-
-				// Setup test module
-				const testNotifierConfig: NotifierConfig = new NotifierConfig();
-				beforeEachWithConfig( testNotifierConfig );
-				componentFixture.detectChanges();
-				spyOn( queueService, 'continue' );
-
-				// Show first notification
-				queueService.push( {
-					payload: {
-						message: 'Whut.',
-						type: 'ERROR'
-					},
-					type: 'SHOW'
-				} );
-				componentFixture.detectChanges();
-				const mockFirstNotificationComponent: MockNotifierNotificationComponent = new MockNotifierNotificationComponent();
-				spyOn( mockFirstNotificationComponent, 'shift' ).and.callThrough(); // Continue
-				// tslint:disable no-any
-				componentInstance.onNotificationReady( <any> mockFirstNotificationComponent ); // Trigger the ready event manually
-				// tslint:enable no-any
-
-				// Show second notification
-				const testNotificationId: string = 'FAKE_ID';
-				queueService.push( {
-					payload: {
-						id: testNotificationId,
-						message: 'Lorem ipsum dolor sit amet.',
-						type: 'SUCCESS'
-					},
-					type: 'SHOW'
-				} );
-				componentFixture.detectChanges();
-				const mockSecondNotificationComponent: MockNotifierNotificationComponent = new MockNotifierNotificationComponent();
-				spyOn( mockSecondNotificationComponent, 'hide' ).and.callThrough(); // Continue
-				// tslint:disable no-any
-				componentInstance.onNotificationReady( <any> mockSecondNotificationComponent ); // Trigger the ready event manually
-				// tslint:enable no-any
-
-				// Hide second notification
-				queueService.push( {
-					payload: testNotificationId,
-					type: 'HIDE'
-				} );
-				componentFixture.detectChanges();
-
-				tick( testNotifierConfig.animations.hide.speed - <number> testNotifierConfig.animations.overlap );
-				componentFixture.detectChanges(); // Run a second change detection (to update the template)
-
-				const listElements: Array<DebugElement> = componentFixture.debugElement.queryAll( By.css( '.x-notifier__container-list' ) );
-
-				const expectedCallTimes: number = 3;
-				expect( listElements.length ).toBe( 1 );
-				expect( mockFirstNotificationComponent.shift ).toHaveBeenCalled();
-				expect( mockSecondNotificationComponent.hide ).toHaveBeenCalled();
-				expect( queueService.continue ).toHaveBeenCalledTimes( expectedCallTimes );
-
-			} ) );
-
-			it( 'should hide the notification when the dismiss event gets dispatched', fakeAsync( () => {
-
-				// Setup test module
-				beforeEachWithConfig( new NotifierConfig( {
-					animations: {
-						enabled: false
-					}
-				} ) );
-				componentFixture.detectChanges();
-				spyOn( queueService, 'continue' );
-
-				// Show second notification
-				const testNotificationId: string = 'FAKE_ID';
-				queueService.push( {
-					payload: {
-						id: testNotificationId,
-						message: 'Lorem ipsum dolor sit amet.',
-						type: 'SUCCESS'
-					},
-					type: 'SHOW'
-				} );
-				componentFixture.detectChanges();
-				const mockNotificationComponent: MockNotifierNotificationComponent = new MockNotifierNotificationComponent();
-				spyOn( mockNotificationComponent, 'hide' ).and.callThrough(); // Continue
-				// tslint:disable no-any
-				componentInstance.onNotificationReady( <any> mockNotificationComponent ); // Trigger the ready event manually
-				// tslint:enable no-any
-
-				// Hide second notification
-				componentInstance.onNotificationDismiss( testNotificationId );
-				tick();
-				componentFixture.detectChanges();
-
-				const listElements: Array<DebugElement> = componentFixture.debugElement.queryAll( By.css( '.x-notifier__container-list' ) );
-
-				const expectedCallTimes: number = 2;
-				expect( listElements.length ).toBe( 0 );
-				expect( mockNotificationComponent.hide ).toHaveBeenCalled();
-				expect( queueService.continue ).toHaveBeenCalledTimes( expectedCallTimes );
-
-			} ) );
-
-		} );
-
-		describe( '(hide oldest / newest)', () => {
-
-			it( 'should hide the oldest notification', fakeAsync( () => {
-
-				// Setup test module
-				beforeEachWithConfig( new NotifierConfig( {
-					animations: {
-						enabled: false
-					}
-				} ) );
-				componentFixture.detectChanges();
-				spyOn( queueService, 'continue' );
-
-				// Show first notification
-				queueService.push( {
-					payload: {
-						id: 'FAKE_ID',
-						message: 'Lorem ipsum dolor sit amet.',
-						type: 'SUCCESS'
-					},
-					type: 'SHOW'
-				} );
-				componentFixture.detectChanges();
-				const mockFirstNotificationComponent: MockNotifierNotificationComponent = new MockNotifierNotificationComponent();
-				spyOn( mockFirstNotificationComponent, 'hide' ).and.callThrough(); // Continue
-				// tslint:disable no-any
-				componentInstance.onNotificationReady( <any> mockFirstNotificationComponent ); // Trigger the ready event manually
-				// tslint:enable no-any
-
-				// Show first notification
-				queueService.push( {
-					payload: {
-						message: 'Whut.',
-						type: 'ERROR'
-					},
-					type: 'SHOW'
-				} );
-				componentFixture.detectChanges();
-				const mockSecondNotificationComponent: MockNotifierNotificationComponent = new MockNotifierNotificationComponent();
-				spyOn( mockSecondNotificationComponent, 'hide' ).and.callThrough(); // Continue
-				// tslint:disable no-any
-				componentInstance.onNotificationReady( <any> mockSecondNotificationComponent ); // Trigger the ready event manually
-				// tslint:enable no-any
-
-				// Hide second notification
-				queueService.push( {
-					type: 'HIDE_OLDEST'
-				} );
-				componentFixture.detectChanges();
-
-				tick();
-				componentFixture.detectChanges(); // Run a second change detection (to update the template)
-
-				const listElements: Array<DebugElement> = componentFixture.debugElement.queryAll( By.css( '.x-notifier__container-list' ) );
-
-				const expectedCallTimes: number = 3;
-				expect( listElements.length ).toBe( 1 );
-				expect( mockFirstNotificationComponent.hide ).toHaveBeenCalled();
-				expect( mockSecondNotificationComponent.hide ).not.toHaveBeenCalled();
-				expect( queueService.continue ).toHaveBeenCalledTimes( expectedCallTimes );
-
-			} ) );
-
-			it( 'should skip hiding the oldest notification if there are no notifications', fakeAsync( () => {
-
-				// Setup test module
-				beforeEachWithConfig( new NotifierConfig( {
-					animations: {
-						enabled: false
-					}
-				} ) );
-				componentFixture.detectChanges();
-				spyOn( queueService, 'continue' );
-
-				// Hide notification
-				queueService.push( {
-					type: 'HIDE_OLDEST'
-				} );
-				componentFixture.detectChanges();
-				tick();
-
-				expect( queueService.continue ).toHaveBeenCalled();
-
-			} ) );
-
-			it( 'should hide the newest notification', fakeAsync( () => {
-
-				// Setup test module
-				beforeEachWithConfig( new NotifierConfig( {
-					animations: {
-						enabled: false
-					}
-				} ) );
-				componentFixture.detectChanges();
-				spyOn( queueService, 'continue' );
-
-				// Show first notification
-				queueService.push( {
-					payload: {
-						message: 'Whut.',
-						type: 'ERROR'
-					},
-					type: 'SHOW'
-				} );
-				componentFixture.detectChanges();
-				const mockFirstNotificationComponent: MockNotifierNotificationComponent = new MockNotifierNotificationComponent();
-				spyOn( mockFirstNotificationComponent, 'hide' ).and.callThrough(); // Continue
-				// tslint:disable no-any
-				componentInstance.onNotificationReady( <any> mockFirstNotificationComponent ); // Trigger the ready event manually
-				// tslint:enable no-any
-
-				// Show first notification
-				queueService.push( {
-					payload: {
-						id: 'FAKE_ID',
-						message: 'Lorem ipsum dolor sit amet.',
-						type: 'SUCCESS'
-					},
-					type: 'SHOW'
-				} );
-				componentFixture.detectChanges();
-				const mockSecondNotificationComponent: MockNotifierNotificationComponent = new MockNotifierNotificationComponent();
-				spyOn( mockSecondNotificationComponent, 'hide' ).and.callThrough(); // Continue
-				// tslint:disable no-any
-				componentInstance.onNotificationReady( <any> mockSecondNotificationComponent ); // Trigger the ready event manually
-				// tslint:enable no-any
-
-				// Hide second notification
-				queueService.push( {
-					type: 'HIDE_NEWEST'
-				} );
-				componentFixture.detectChanges();
-
-				tick();
-				componentFixture.detectChanges(); // Run a second change detection (to update the template)
-
-				const listElements: Array<DebugElement> = componentFixture.debugElement.queryAll( By.css( '.x-notifier__container-list' ) );
-
-				const expectedCallTimes: number = 3;
-				expect( listElements.length ).toBe( 1 );
-				expect( mockFirstNotificationComponent.hide ).not.toHaveBeenCalled();
-				expect( mockSecondNotificationComponent.hide ).toHaveBeenCalled();
-				expect( queueService.continue ).toHaveBeenCalledTimes( expectedCallTimes );
-
-			} ) );
-
-			it( 'should skip hiding the newest notification if there are no notifications', fakeAsync( () => {
-
-				// Setup test module
-				beforeEachWithConfig( new NotifierConfig( {
-					animations: {
-						enabled: false
-					}
-				} ) );
-				componentFixture.detectChanges();
-				spyOn( queueService, 'continue' );
-
-				// Hide notification
-				queueService.push( {
-					type: 'HIDE_NEWEST'
-				} );
-				componentFixture.detectChanges();
-				tick();
-
-				expect( queueService.continue ).toHaveBeenCalled();
-
-			} ) );
-
-		} );
-
-		describe( '(hide all)', () => {
-
-			it( 'should hide all notifications', fakeAsync( () => {
-
-				// Setup test module
-				beforeEachWithConfig( new NotifierConfig( {
-					animations: {
-						enabled: false
-					}
-				} ) );
-				componentFixture.detectChanges();
-				spyOn( queueService, 'continue' );
-
-				// Show first notification
-				queueService.push( {
-					payload: {
-						message: 'Lorem ipsum dolor sit amet.',
-						type: 'SUCCESS'
-					},
-					type: 'SHOW'
-				} );
-				componentFixture.detectChanges();
-				const mockFirstNotificationComponent: MockNotifierNotificationComponent = new MockNotifierNotificationComponent();
-				spyOn( mockFirstNotificationComponent, 'hide' ).and.callThrough(); // Continue
-				// tslint:disable no-any
-				componentInstance.onNotificationReady( <any> mockFirstNotificationComponent ); // Trigger the ready event manually
-				// tslint:enable no-any
-
-				// Show first notification
-				queueService.push( {
-					payload: {
-						message: 'Whut.',
-						type: 'ERROR'
-					},
-					type: 'SHOW'
-				} );
-				componentFixture.detectChanges();
-				const mockSecondNotificationComponent: MockNotifierNotificationComponent = new MockNotifierNotificationComponent();
-				spyOn( mockSecondNotificationComponent, 'hide' ).and.callThrough(); // Continue
-				// tslint:disable no-any
-				componentInstance.onNotificationReady( <any> mockSecondNotificationComponent ); // Trigger the ready event manually
-				// tslint:enable no-any
-
-				// Hide second notification
-				queueService.push( {
-					type: 'HIDE_ALL'
-				} );
-				componentFixture.detectChanges();
-
-				tick();
-				componentFixture.detectChanges(); // Run a second change detection (to update the template)
-
-				const listElements: Array<DebugElement> = componentFixture.debugElement.queryAll( By.css( '.x-notifier__container-list' ) );
-
-				const expectedCallTimes: number = 3;
-				expect( listElements.length ).toBe( 0 );
-				expect( mockFirstNotificationComponent.hide ).toHaveBeenCalled();
-				expect( mockSecondNotificationComponent.hide ).toHaveBeenCalled();
-				expect( queueService.continue ).toHaveBeenCalledTimes( expectedCallTimes );
-
-			} ) );
-
-			it( 'should hide all notifications (with animations)', fakeAsync( () => {
-
-				// Setup test module
-				const testNotifierConfig: NotifierConfig = new NotifierConfig();
-				beforeEachWithConfig( testNotifierConfig );
-				componentFixture.detectChanges();
-				spyOn( queueService, 'continue' );
-
-				// Show first notification
-				queueService.push( {
-					payload: {
-						message: 'Lorem ipsum dolor sit amet.',
-						type: 'SUCCESS'
-					},
-					type: 'SHOW'
-				} );
-				componentFixture.detectChanges();
-				const mockFirstNotificationComponent: MockNotifierNotificationComponent = new MockNotifierNotificationComponent();
-				spyOn( mockFirstNotificationComponent, 'hide' ).and.callThrough(); // Continue
-				// tslint:disable no-any
-				componentInstance.onNotificationReady( <any> mockFirstNotificationComponent ); // Trigger the ready event manually
-				// tslint:enable no-any
-
-				// Show first notification
-				queueService.push( {
-					payload: {
-						message: 'Whut.',
-						type: 'ERROR'
-					},
-					type: 'SHOW'
-				} );
-				componentFixture.detectChanges();
-				const mockSecondNotificationComponent: MockNotifierNotificationComponent = new MockNotifierNotificationComponent();
-				spyOn( mockSecondNotificationComponent, 'hide' ).and.callThrough(); // Continue
-				// tslint:disable no-any
-				componentInstance.onNotificationReady( <any> mockSecondNotificationComponent ); // Trigger the ready event manually
-				// tslint:enable no-any
-
-				// Hide second notification
-				queueService.push( {
-					type: 'HIDE_ALL'
-				} );
-				componentFixture.detectChanges();
-
-				const numberOfNotifications: number = 2;
-				tick( testNotifierConfig.animations.hide.speed + ( numberOfNotifications * <number> testNotifierConfig.animations.hide.offset ) );
-				componentFixture.detectChanges(); // Run a second change detection (to update the template)
-
-				const listElements: Array<DebugElement> = componentFixture.debugElement.queryAll( By.css( '.x-notifier__container-list' ) );
-
-				const expectedCallTimes: number = 3;
-				expect( listElements.length ).toBe( 0 );
-				expect( mockFirstNotificationComponent.hide ).toHaveBeenCalled();
-				expect( mockSecondNotificationComponent.hide ).toHaveBeenCalled();
-				expect( queueService.continue ).toHaveBeenCalledTimes( expectedCallTimes );
-
-			} ) );
-
-			it( 'should skip hiding all notification if there are no notifications', fakeAsync( () => {
-
-				// Setup test module
-				beforeEachWithConfig( new NotifierConfig( {
-					animations: {
-						enabled: false
-					}
-				} ) );
-				componentFixture.detectChanges();
-				spyOn( queueService, 'continue' );
-
-				// Hide notification
-				queueService.push( {
-					type: 'HIDE_ALL'
-				} );
-				componentFixture.detectChanges();
-				tick();
-
-				expect( queueService.continue ).toHaveBeenCalled();
-
-			} ) );
-
-		} );
-
-		/**
-		 * Helper for upfront configuration
-		 */
-		function beforeEachWithConfig( testNotifierConfig: NotifierConfig ): void {
-
-			TestBed.configureTestingModule( {
-				declarations: [
-					NotifierContainerComponent
-				],
-				providers: [
-					{
-						provide: NotifierService,
-						useValue: {
-							getConfig: () => testNotifierConfig
-						}
-					},
-					{
-						provide: NotifierQueueService,
-						useClass: MockNotifierQueueService
-					}
-				],
-				schemas: [
-					NO_ERRORS_SCHEMA // Ignore sub-components (inknown tags in the HTML template)
-				]
+			componentFixture.detectChanges();
+			jest.spyOn( queueService, 'continue' );
+
+			// Show first notification
+			queueService.push( {
+				payload: {
+					message: 'Whut.',
+					type: 'ERROR'
+				},
+				type: 'SHOW'
 			} );
-			componentFixture = TestBed.createComponent( NotifierContainerComponent );
-			componentInstance = componentFixture.componentInstance;
+			componentFixture.detectChanges();
+			const mockFirstNotificationComponent: MockNotifierNotificationComponent = new MockNotifierNotificationComponent();
+			jest.spyOn( mockFirstNotificationComponent, 'hide' ); // Continue
+			// tslint:disable no-any
+			componentInstance.onNotificationReady( <any> mockFirstNotificationComponent ); // Trigger the ready event manually
+			// tslint:enable no-any
 
-			queueService = TestBed.get( NotifierQueueService );
+			// Show first notification
+			queueService.push( {
+				payload: {
+					id: 'FAKE_ID',
+					message: 'Lorem ipsum dolor sit amet.',
+					type: 'SUCCESS'
+				},
+				type: 'SHOW'
+			} );
+			componentFixture.detectChanges();
+			const mockSecondNotificationComponent: MockNotifierNotificationComponent = new MockNotifierNotificationComponent();
+			jest.spyOn( mockSecondNotificationComponent, 'hide' ); // Continue
+			// tslint:disable no-any
+			componentInstance.onNotificationReady( <any> mockSecondNotificationComponent ); // Trigger the ready event manually
+			// tslint:enable no-any
 
-		}
+			// Hide second notification
+			queueService.push( {
+				type: 'HIDE_NEWEST'
+			} );
+			componentFixture.detectChanges();
+
+			tick();
+			componentFixture.detectChanges(); // Run a second change detection (to update the template)
+
+			const listElements: Array<DebugElement> = componentFixture.debugElement.queryAll( By.css( '.x-notifier__container-list' ) );
+
+			const expectedCallTimes: number = 3;
+			expect( listElements.length ).toBe( 1 );
+			expect( mockFirstNotificationComponent.hide ).not.toHaveBeenCalled();
+			expect( mockSecondNotificationComponent.hide ).toHaveBeenCalled();
+			expect( queueService.continue ).toHaveBeenCalledTimes( expectedCallTimes );
+
+		} ) );
+
+		it( 'should skip hiding the newest notification if there are no notifications', fakeAsync( () => {
+
+			// Setup test module
+			beforeEachWithConfig( new NotifierConfig( {
+				animations: {
+					enabled: false
+				}
+			} ) );
+			componentFixture.detectChanges();
+			jest.spyOn( queueService, 'continue' );
+
+			// Hide notification
+			queueService.push( {
+				type: 'HIDE_NEWEST'
+			} );
+			componentFixture.detectChanges();
+			tick();
+
+			expect( queueService.continue ).toHaveBeenCalled();
+
+		} ) );
 
 	} );
 
-}
+	describe( '(hide all)', () => {
+
+		it( 'should hide all notifications', fakeAsync( () => {
+
+			// Setup test module
+			beforeEachWithConfig( new NotifierConfig( {
+				animations: {
+					enabled: false
+				}
+			} ) );
+			componentFixture.detectChanges();
+			jest.spyOn( queueService, 'continue' );
+
+			// Show first notification
+			queueService.push( {
+				payload: {
+					message: 'Lorem ipsum dolor sit amet.',
+					type: 'SUCCESS'
+				},
+				type: 'SHOW'
+			} );
+			componentFixture.detectChanges();
+			const mockFirstNotificationComponent: MockNotifierNotificationComponent = new MockNotifierNotificationComponent();
+			jest.spyOn( mockFirstNotificationComponent, 'hide' ); // Continue
+			// tslint:disable no-any
+			componentInstance.onNotificationReady( <any> mockFirstNotificationComponent ); // Trigger the ready event manually
+			// tslint:enable no-any
+
+			// Show first notification
+			queueService.push( {
+				payload: {
+					message: 'Whut.',
+					type: 'ERROR'
+				},
+				type: 'SHOW'
+			} );
+			componentFixture.detectChanges();
+			const mockSecondNotificationComponent: MockNotifierNotificationComponent = new MockNotifierNotificationComponent();
+			jest.spyOn( mockSecondNotificationComponent, 'hide' ); // Continue
+			// tslint:disable no-any
+			componentInstance.onNotificationReady( <any> mockSecondNotificationComponent ); // Trigger the ready event manually
+			// tslint:enable no-any
+
+			// Hide second notification
+			queueService.push( {
+				type: 'HIDE_ALL'
+			} );
+			componentFixture.detectChanges();
+
+			tick();
+			componentFixture.detectChanges(); // Run a second change detection (to update the template)
+
+			const listElements: Array<DebugElement> = componentFixture.debugElement.queryAll( By.css( '.x-notifier__container-list' ) );
+
+			const expectedCallTimes: number = 3;
+			expect( listElements.length ).toBe( 0 );
+			expect( mockFirstNotificationComponent.hide ).toHaveBeenCalled();
+			expect( mockSecondNotificationComponent.hide ).toHaveBeenCalled();
+			expect( queueService.continue ).toHaveBeenCalledTimes( expectedCallTimes );
+
+		} ) );
+
+		it( 'should hide all notifications (with animations)', fakeAsync( () => {
+
+			// Setup test module
+			const testNotifierConfig: NotifierConfig = new NotifierConfig();
+			beforeEachWithConfig( testNotifierConfig );
+			componentFixture.detectChanges();
+			jest.spyOn( queueService, 'continue' );
+
+			// Show first notification
+			queueService.push( {
+				payload: {
+					message: 'Lorem ipsum dolor sit amet.',
+					type: 'SUCCESS'
+				},
+				type: 'SHOW'
+			} );
+			componentFixture.detectChanges();
+			const mockFirstNotificationComponent: MockNotifierNotificationComponent = new MockNotifierNotificationComponent();
+			jest.spyOn( mockFirstNotificationComponent, 'hide' ); // Continue
+			// tslint:disable no-any
+			componentInstance.onNotificationReady( <any> mockFirstNotificationComponent ); // Trigger the ready event manually
+			// tslint:enable no-any
+
+			// Show first notification
+			queueService.push( {
+				payload: {
+					message: 'Whut.',
+					type: 'ERROR'
+				},
+				type: 'SHOW'
+			} );
+			componentFixture.detectChanges();
+			const mockSecondNotificationComponent: MockNotifierNotificationComponent = new MockNotifierNotificationComponent();
+			jest.spyOn( mockSecondNotificationComponent, 'hide' ); // Continue
+			// tslint:disable no-any
+			componentInstance.onNotificationReady( <any> mockSecondNotificationComponent ); // Trigger the ready event manually
+			// tslint:enable no-any
+
+			// Hide second notification
+			queueService.push( {
+				type: 'HIDE_ALL'
+			} );
+			componentFixture.detectChanges();
+
+			const numberOfNotifications: number = 2;
+			tick( testNotifierConfig.animations.hide.speed + ( numberOfNotifications * <number> testNotifierConfig.animations.hide.offset ) );
+			componentFixture.detectChanges(); // Run a second change detection (to update the template)
+
+			const listElements: Array<DebugElement> = componentFixture.debugElement.queryAll( By.css( '.x-notifier__container-list' ) );
+
+			const expectedCallTimes: number = 3;
+			expect( listElements.length ).toBe( 0 );
+			expect( mockFirstNotificationComponent.hide ).toHaveBeenCalled();
+			expect( mockSecondNotificationComponent.hide ).toHaveBeenCalled();
+			expect( queueService.continue ).toHaveBeenCalledTimes( expectedCallTimes );
+
+		} ) );
+
+		it( 'should skip hiding all notification if there are no notifications', fakeAsync( () => {
+
+			// Setup test module
+			beforeEachWithConfig( new NotifierConfig( {
+				animations: {
+					enabled: false
+				}
+			} ) );
+			componentFixture.detectChanges();
+			jest.spyOn( queueService, 'continue' );
+
+			// Hide notification
+			queueService.push( {
+				type: 'HIDE_ALL'
+			} );
+			componentFixture.detectChanges();
+			tick();
+
+			expect( queueService.continue ).toHaveBeenCalled();
+
+		} ) );
+
+	} );
+
+	/**
+	 * Helper for upfront configuration
+	 */
+	function beforeEachWithConfig( testNotifierConfig: NotifierConfig ): void {
+
+		TestBed.configureTestingModule( {
+			declarations: [
+				NotifierContainerComponent
+			],
+			providers: [
+				{
+					provide: NotifierService,
+					useValue: {
+						getConfig: () => testNotifierConfig
+					}
+				},
+				{
+					provide: NotifierQueueService,
+					useClass: MockNotifierQueueService
+				}
+			],
+			schemas: [
+				NO_ERRORS_SCHEMA // Ignore sub-components (inknown tags in the HTML template)
+			]
+		} );
+		componentFixture = TestBed.createComponent( NotifierContainerComponent );
+		componentInstance = componentFixture.componentInstance;
+
+		queueService = TestBed.get( NotifierQueueService );
+
+	}
+
+} );
 
 /**
  * Mock Notifier Queue Service
