+++ src/lib/src/components/notifier-notification.component.spec.ts
@@ -13,964 +13,960 @@ import { NotifierTimerService } from '../services/notifier-timer.service';
 /**
  * Notifier Notification Component - Unit Test
  */
-export function main(): void {
+describe( 'Notifier Notification Component', () => {
 
-	describe( 'Notifier Notification Component', () => {
+	const testNotification: NotifierNotification = new NotifierNotification( {
+		id: 'ID_FAKE',
+		message: 'Lorem ipsum dolor sit amet.',
+		type: 'SUCCESS'
+	} );
 
-		const testNotification: NotifierNotification = new NotifierNotification( {
-			id: 'ID_FAKE',
-			message: 'Lorem ipsum dolor sit amet.',
-			type: 'SUCCESS'
-		} );
+	let componentFixture: ComponentFixture<NotifierNotificationComponent>;
+	let componentInstance: NotifierNotificationComponent;
 
-		let componentFixture: ComponentFixture<NotifierNotificationComponent>;
-		let componentInstance: NotifierNotificationComponent;
+	let timerService: MockNotifierTimerService;
 
-		let timerService: MockNotifierTimerService;
+	it( 'should instantiate', () => {
 
-		it( 'should instantiate', () => {
+		// Setup test module
+		beforeEachWithConfig( new NotifierConfig() );
+
+		expect( componentInstance ).toBeDefined();
+
+	} );
+
+	describe( '(render)', () => {
+
+		it( 'should render', () => {
 
 			// Setup test module
 			beforeEachWithConfig( new NotifierConfig() );
 
-			expect( componentInstance ).toBeDefined();
+			componentInstance.notification = testNotification;
+			componentFixture.detectChanges();
+
+			// Check the calculated values
+			expect( componentInstance.getConfig() ).toEqual( new NotifierConfig() );
+			expect( componentInstance.getHeight() ).toBe( componentFixture.nativeElement.offsetHeight );
+			expect( componentInstance.getWidth() ).toBe( componentFixture.nativeElement.offsetWidth );
+			expect( componentInstance.getShift() ).toBe( 0 );
+
+			// Check the template
+			const messageElement: DebugElement = componentFixture.debugElement.query( By.css( '.x-notifier__notification-message' ) );
+			expect( messageElement.nativeElement.textContent ).toContain( componentInstance.notification.message );
+			const dismissButtonElement: DebugElement = componentFixture.debugElement.query( By.css( '.x-notifier__notification-button' ) );
+			expect( dismissButtonElement ).not.toBeNull();
+
+			// Check the class names
+			const classNameType: string = `x-notifier__notification--${ componentInstance.notification.type }`;
+			expect( componentFixture.nativeElement.classList.contains( classNameType ) ).toBeTruthy();
+			const classNameTheme: string = `x-notifier__notification--${ componentInstance.getConfig().theme }`;
+			expect( componentFixture.nativeElement.classList.contains( classNameTheme ) ).toBeTruthy();
 
 		} );
 
-		describe( '(render)', () => {
+		it( 'should render on the left', () => {
 
-			it( 'should render', () => {
-
-				// Setup test module
-				beforeEachWithConfig( new NotifierConfig() );
-
-				componentInstance.notification = testNotification;
-				componentFixture.detectChanges();
-
-				// Check the calculated values
-				expect( componentInstance.getConfig() ).toEqual( new NotifierConfig() );
-				expect( componentInstance.getHeight() ).toBe( componentFixture.nativeElement.offsetHeight );
-				expect( componentInstance.getWidth() ).toBe( componentFixture.nativeElement.offsetWidth );
-				expect( componentInstance.getShift() ).toBe( 0 );
-
-				// Check the template
-				const messageElement: DebugElement = componentFixture.debugElement.query( By.css( '.x-notifier__notification-message' ) );
-				expect( messageElement.nativeElement.textContent ).toContain( componentInstance.notification.message );
-				const dismissButtonElement: DebugElement = componentFixture.debugElement.query( By.css( '.x-notifier__notification-button' ) );
-				expect( dismissButtonElement ).not.toBeNull();
-
-				// Check the class names
-				const classNameType: string = `x-notifier__notification--${ componentInstance.notification.type }`;
-				expect( componentFixture.nativeElement.classList.contains( classNameType ) ).toBeTruthy();
-				const classNameTheme: string = `x-notifier__notification--${ componentInstance.getConfig().theme }`;
-				expect( componentFixture.nativeElement.classList.contains( classNameTheme ) ).toBeTruthy();
-
-			} );
-
-			it( 'should render on the left', () => {
-
-				// Setup test module
-				const testNotifierConfig: NotifierConfig = new NotifierConfig( {
-					position: {
-						horizontal: {
-							distance: 10,
-							position: 'left'
-						},
-						vertical: {
-							distance: 10,
-							gap: 4,
-							position: 'top'
-						}
+			// Setup test module
+			const testNotifierConfig: NotifierConfig = new NotifierConfig( {
+				position: {
+					horizontal: {
+						distance: 10,
+						position: 'left'
+					},
+					vertical: {
+						distance: 10,
+						gap: 4,
+						position: 'top'
 					}
-				} );
-				beforeEachWithConfig( testNotifierConfig );
-
-				componentInstance.notification = testNotification;
-				componentFixture.detectChanges();
-
-				// Check position
-				expect( componentFixture.debugElement.styles[ 'left' ] ).toBe( `${ testNotifierConfig.position.horizontal.distance }px` );
-
+				}
 			} );
+			beforeEachWithConfig( testNotifierConfig );
 
-			it( 'should render on the right', () => {
+			componentInstance.notification = testNotification;
+			componentFixture.detectChanges();
 
-				// Setup test module
-				const testNotifierConfig: NotifierConfig = new NotifierConfig( {
-					position: {
-						horizontal: {
-							distance: 10,
-							position: 'right'
-						},
-						vertical: {
-							distance: 10,
-							gap: 4,
-							position: 'top'
-						}
-					}
-				} );
-				beforeEachWithConfig( testNotifierConfig );
-
-				componentInstance.notification = testNotification;
-				componentFixture.detectChanges();
-
-				// Check position
-				expect( componentFixture.debugElement.styles[ 'right' ] ).toBe( `${ testNotifierConfig.position.horizontal.distance }px` );
-
-			} );
-
-			it( 'should render in the middle', () => {
-
-				// Setup test module
-				const testNotifierConfig: NotifierConfig = new NotifierConfig( {
-					position: {
-						horizontal: {
-							distance: 10,
-							position: 'middle'
-						},
-						vertical: {
-							distance: 10,
-							gap: 4,
-							position: 'top'
-						}
-					}
-				} );
-				beforeEachWithConfig( testNotifierConfig );
-
-				componentInstance.notification = testNotification;
-				componentFixture.detectChanges();
-
-				// Check position
-				expect( componentFixture.debugElement.styles[ 'left' ] ).toBe( '50%' );
-				expect( componentFixture.debugElement.styles[ 'transform' ] ).toBe( 'translate3d( -50%, 0, 0 )' );
-
-			} );
-
-			it( 'should render on the top', () => {
-
-				// Setup test module
-				const testNotifierConfig: NotifierConfig = new NotifierConfig( {
-					position: {
-						horizontal: {
-							distance: 10,
-							position: 'left'
-						},
-						vertical: {
-							distance: 10,
-							gap: 4,
-							position: 'top'
-						}
-					}
-				} );
-				beforeEachWithConfig( testNotifierConfig );
-
-				componentInstance.notification = testNotification;
-				componentFixture.detectChanges();
-
-				// Check position
-				expect( componentFixture.debugElement.styles[ 'top' ] ).toBe( `${ testNotifierConfig.position.vertical.distance }px` );
-
-			} );
-
-			it( 'should render on the bottom', () => {
-
-				// Setup test module
-				const testNotifierConfig: NotifierConfig = new NotifierConfig( {
-					position: {
-						horizontal: {
-							distance: 10,
-							position: 'left'
-						},
-						vertical: {
-							distance: 10,
-							gap: 4,
-							position: 'bottom'
-						}
-					}
-				} );
-				beforeEachWithConfig( testNotifierConfig );
-
-				componentInstance.notification = testNotification;
-				componentFixture.detectChanges();
-
-				// Check position
-				expect( componentFixture.debugElement.styles[ 'bottom' ] ).toBe( `${ testNotifierConfig.position.vertical.distance }px` );
-
-			} );
+			// Check position
+			expect( componentFixture.debugElement.styles[ 'left' ] ).toBe( `${ testNotifierConfig.position.horizontal.distance }px` );
 
 		} );
 
-		describe( '(show)', () => {
+		it( 'should render on the right', () => {
 
-			it( 'should show', fakeAsync( () => {
-
-				// Setup test module
-				beforeEachWithConfig( new NotifierConfig( {
-					animations: {
-						enabled: false
+			// Setup test module
+			const testNotifierConfig: NotifierConfig = new NotifierConfig( {
+				position: {
+					horizontal: {
+						distance: 10,
+						position: 'right'
 					},
-					behaviour: {
-						autoHide: false
+					vertical: {
+						distance: 10,
+						gap: 4,
+						position: 'top'
 					}
-				} ) );
+				}
+			} );
+			beforeEachWithConfig( testNotifierConfig );
 
-				componentInstance.notification = testNotification;
-				componentFixture.detectChanges();
+			componentInstance.notification = testNotification;
+			componentFixture.detectChanges();
 
-				const showCallback: () => {} = jasmine.createSpy( 'showCallback' );
-				componentInstance.show().then( showCallback );
-				tick();
-
-				expect( componentFixture.debugElement.styles[ 'visibility' ] ).toBe( 'visible' );
-				expect( showCallback ).toHaveBeenCalled();
-
-			} ) );
-
-			it( 'should show (with animations)', fakeAsync( () => {
-
-				// Setup test module
-				beforeEachWithConfig( new NotifierConfig( {
-					behaviour: {
-						autoHide: false
-					}
-				} ) );
-
-				componentInstance.notification = testNotification;
-				componentFixture.detectChanges();
-
-				// Mock away the Web Animations API
-				spyOn( componentFixture.nativeElement, 'animate' ).and.callFake( () => {
-					return {
-						finished: new Promise<undefined>( ( resolve: () => void, reject: () => void ) => {
-							componentFixture.debugElement.styles[ 'opacity' ] = '1'; // Fake animation result
-							resolve();
-						} )
-					};
-				} );
-
-				const showCallback: () => {} = jasmine.createSpy( 'showCallback' );
-				componentInstance.show().then( showCallback );
-				tick();
-
-				expect( componentFixture.debugElement.styles[ 'visibility' ] ).toBe( 'visible' );
-				expect( componentFixture.debugElement.styles[ 'opacity' ] ).toBe( '1' );
-				expect( showCallback ).toHaveBeenCalled();
-
-			} ) );
+			// Check position
+			expect( componentFixture.debugElement.styles[ 'right' ] ).toBe( `${ testNotifierConfig.position.horizontal.distance }px` );
 
 		} );
 
-		describe( '(hide)', () => {
+		it( 'should render in the middle', () => {
 
-			it( 'should hide', fakeAsync( () => {
-
-				// Setup test module
-				beforeEachWithConfig( new NotifierConfig( {
-					animations: {
-						enabled: false
+			// Setup test module
+			const testNotifierConfig: NotifierConfig = new NotifierConfig( {
+				position: {
+					horizontal: {
+						distance: 10,
+						position: 'middle'
 					},
-					behaviour: {
-						autoHide: false
+					vertical: {
+						distance: 10,
+						gap: 4,
+						position: 'top'
 					}
-				} ) );
+				}
+			} );
+			beforeEachWithConfig( testNotifierConfig );
 
-				componentInstance.notification = testNotification;
-				componentFixture.detectChanges();
+			componentInstance.notification = testNotification;
+			componentFixture.detectChanges();
 
-				const hideCallback: () => {} = jasmine.createSpy( 'hideCallback' );
-				componentInstance.hide().then( hideCallback );
-				tick();
-
-				expect( hideCallback ).toHaveBeenCalled();
-
-			} ) );
-
-			it( 'should hide (with animations)', fakeAsync( () => {
-
-				// Setup test module
-				beforeEachWithConfig( new NotifierConfig( {
-					behaviour: {
-						autoHide: false
-					}
-				} ) );
-
-				componentInstance.notification = testNotification;
-				componentFixture.detectChanges();
-
-				// Mock away the Web Animations API
-				spyOn( componentFixture.nativeElement, 'animate' ).and.callFake( () => {
-					return {
-						finished: new Promise<undefined>( ( resolve: () => void, reject: () => void ) => {
-							componentFixture.debugElement.styles[ 'opacity' ] = '0'; // Fake animation result
-							resolve();
-						} )
-					};
-				} );
-
-				const hideCallback: () => {} = jasmine.createSpy( 'hideCallback' );
-				componentInstance.hide().then( hideCallback );
-				tick();
-
-				expect( componentFixture.debugElement.styles[ 'opacity' ] ).toBe( '0' );
-				expect( hideCallback ).toHaveBeenCalled();
-
-			} ) );
+			// Check position
+			expect( componentFixture.debugElement.styles[ 'left' ] ).toBe( '50%' );
+			expect( componentFixture.debugElement.styles[ 'transform' ] ).toBe( 'translate3d( -50%, 0, 0 )' );
 
 		} );
 
-		describe( '(shift)', () => {
+		it( 'should render on the top', () => {
 
-			it( 'should shift to make place on top', fakeAsync( () => {
-
-				// Setup test module
-				const testNotifierConfig: NotifierConfig = new NotifierConfig( {
-					animations: {
-						enabled: false
+			// Setup test module
+			const testNotifierConfig: NotifierConfig = new NotifierConfig( {
+				position: {
+					horizontal: {
+						distance: 10,
+						position: 'left'
 					},
-					behaviour: {
-						autoHide: false
-					},
-					position: {
-						horizontal: {
-							distance: 12,
-							position: 'left'
-						},
-						vertical: {
-							distance: 12,
-							gap: 10,
-							position: 'top'
-						}
+					vertical: {
+						distance: 10,
+						gap: 4,
+						position: 'top'
 					}
-				} );
-				beforeEachWithConfig( testNotifierConfig );
-
-				componentInstance.notification = testNotification;
-				componentFixture.detectChanges();
-
-				const shiftCallback: () => {} = jasmine.createSpy( 'shiftCallback' );
-				const shiftDistance: number = 100;
-				componentInstance.shift( shiftDistance, true ).then( shiftCallback );
-				tick();
-
-				expect( componentFixture.debugElement.styles[ 'transform' ] )
-					.toBe( `translate3d( 0, ${ shiftDistance + testNotifierConfig.position.vertical.gap }px, 0 )` );
-				expect( shiftCallback ).toHaveBeenCalled();
-
-			} ) );
-
-			it( 'should shift to make place on top (with animations)', fakeAsync( () => {
-
-				// Setup test module
-				const testNotifierConfig: NotifierConfig = new NotifierConfig( {
-					behaviour: {
-						autoHide: false
-					},
-					position: {
-						horizontal: {
-							distance: 12,
-							position: 'left'
-						},
-						vertical: {
-							distance: 12,
-							gap: 10,
-							position: 'top'
-						}
-					}
-				} );
-				beforeEachWithConfig( testNotifierConfig );
-
-				componentInstance.notification = testNotification;
-				componentFixture.detectChanges();
-
-				const shiftDistance: number = 100;
-
-				// Mock away the Web Animations API
-				spyOn( componentFixture.nativeElement, 'animate' ).and.callFake( () => {
-					return {
-						finished: new Promise<undefined>( ( resolve: () => void, reject: () => void ) => {
-							componentFixture.debugElement.styles[ 'transform' ] =
-								`translate3d( 0, ${ shiftDistance + testNotifierConfig.position.vertical.gap }px, 0 )`; // Fake animation result
-							resolve();
-						} )
-					};
-				} );
-
-				const shiftCallback: () => {} = jasmine.createSpy( 'shiftCallback' );
-				componentInstance.shift( shiftDistance, true ).then( shiftCallback );
-				tick();
-
-				expect( componentFixture.debugElement.styles[ 'transform' ] )
-					.toBe( `translate3d( 0, ${ shiftDistance + testNotifierConfig.position.vertical.gap }px, 0 )` );
-				expect( shiftCallback ).toHaveBeenCalled();
-
-			} ) );
-
-			it( 'should shift to make place on bottom', fakeAsync( () => {
-
-				// Setup test module
-				const testNotifierConfig: NotifierConfig = new NotifierConfig( {
-					animations: {
-						enabled: false
-					},
-					behaviour: {
-						autoHide: false
-					},
-					position: {
-						horizontal: {
-							distance: 12,
-							position: 'left'
-						},
-						vertical: {
-							distance: 12,
-							gap: 10,
-							position: 'bottom'
-						}
-					}
-				} );
-				beforeEachWithConfig( testNotifierConfig );
-
-				componentInstance.notification = testNotification;
-				componentFixture.detectChanges();
-
-				const shiftCallback: () => {} = jasmine.createSpy( 'shiftCallback' );
-				const shiftDistance: number = 100;
-				componentInstance.shift( shiftDistance, true ).then( shiftCallback );
-				tick();
-
-				expect( componentFixture.debugElement.styles[ 'transform' ] )
-					.toBe( `translate3d( 0, ${ -shiftDistance - testNotifierConfig.position.vertical.gap }px, 0 )` );
-				expect( shiftCallback ).toHaveBeenCalled();
-
-			} ) );
-
-			it( 'should shift to make place on bottom (with animations)', fakeAsync( () => {
-
-				// Setup test module
-				const testNotifierConfig: NotifierConfig = new NotifierConfig( {
-					behaviour: {
-						autoHide: false
-					},
-					position: {
-						horizontal: {
-							distance: 12,
-							position: 'left'
-						},
-						vertical: {
-							distance: 12,
-							gap: 10,
-							position: 'bottom'
-						}
-					}
-				} );
-				beforeEachWithConfig( testNotifierConfig );
-
-				componentInstance.notification = testNotification;
-				componentFixture.detectChanges();
-
-				// Mock away the Web Animations API
-				const shiftDistance: number = 100;
-				spyOn( componentFixture.nativeElement, 'animate' ).and.callFake( () => {
-					return {
-						finished: new Promise<undefined>( ( resolve: () => void, reject: () => void ) => {
-							componentFixture.debugElement.styles[ 'transform' ] =
-								`translate3d( 0, ${ -shiftDistance - testNotifierConfig.position.vertical.gap }px, 0 )`; // Fake animation result
-							resolve();
-						} )
-					};
-				} );
-
-				const shiftCallback: () => {} = jasmine.createSpy( 'shiftCallback' );
-				componentInstance.shift( shiftDistance, true ).then( shiftCallback );
-				tick();
-
-				expect( componentFixture.debugElement.styles[ 'transform' ] )
-					.toBe( `translate3d( 0, ${ -shiftDistance - testNotifierConfig.position.vertical.gap }px, 0 )` );
-				expect( shiftCallback ).toHaveBeenCalled();
-
-			} ) );
-
-			it( 'should shift to fill place on top', fakeAsync( () => {
-
-				// Setup test module
-				const testNotifierConfig: NotifierConfig = new NotifierConfig( {
-					animations: {
-						enabled: false
-					},
-					behaviour: {
-						autoHide: false
-					},
-					position: {
-						horizontal: {
-							distance: 12,
-							position: 'left'
-						},
-						vertical: {
-							distance: 12,
-							gap: 10,
-							position: 'top'
-						}
-					}
-				} );
-				beforeEachWithConfig( testNotifierConfig );
-
-				componentInstance.notification = testNotification;
-				componentFixture.detectChanges();
-
-				const shiftCallback: () => {} = jasmine.createSpy( 'shiftCallback' );
-				const shiftDistance: number = 100;
-				componentInstance.shift( shiftDistance, false ).then( shiftCallback );
-				tick();
-
-				expect( componentFixture.debugElement.styles[ 'transform' ] )
-					.toBe( `translate3d( 0, ${ -shiftDistance - testNotifierConfig.position.vertical.gap }px, 0 )` );
-				expect( shiftCallback ).toHaveBeenCalled();
-
-			} ) );
-
-			it( 'should shift to fill place on top (with animations)', fakeAsync( () => {
-
-				// Setup test module
-				const testNotifierConfig: NotifierConfig = new NotifierConfig( {
-					behaviour: {
-						autoHide: false
-					},
-					position: {
-						horizontal: {
-							distance: 12,
-							position: 'left'
-						},
-						vertical: {
-							distance: 12,
-							gap: 10,
-							position: 'top'
-						}
-					}
-				} );
-				beforeEachWithConfig( testNotifierConfig );
-
-				componentInstance.notification = testNotification;
-				componentFixture.detectChanges();
-
-				// Mock away the Web Animations API
-				const shiftDistance: number = 100;
-				spyOn( componentFixture.nativeElement, 'animate' ).and.callFake( () => {
-					return {
-						finished: new Promise<undefined>( ( resolve: () => void, reject: () => void ) => {
-							componentFixture.debugElement.styles[ 'transform' ] =
-								`translate3d( 0, ${ -shiftDistance - testNotifierConfig.position.vertical.gap }px, 0 )`; // Fake animation result
-							resolve();
-						} )
-					};
-				} );
-
-				const shiftCallback: () => {} = jasmine.createSpy( 'shiftCallback' );
-				componentInstance.shift( shiftDistance, false ).then( shiftCallback );
-				tick();
-
-				expect( componentFixture.debugElement.styles[ 'transform' ] )
-					.toBe( `translate3d( 0, ${ 0 - shiftDistance - testNotifierConfig.position.vertical.gap }px, 0 )` );
-				expect( shiftCallback ).toHaveBeenCalled();
-
-			} ) );
-
-			it( 'should shift to fill place on bottom', fakeAsync( () => {
-
-				// Setup test module
-				const testNotifierConfig: NotifierConfig = new NotifierConfig( {
-					animations: {
-						enabled: false
-					},
-					behaviour: {
-						autoHide: false
-					},
-					position: {
-						horizontal: {
-							distance: 12,
-							position: 'left'
-						},
-						vertical: {
-							distance: 12,
-							gap: 10,
-							position: 'bottom'
-						}
-					}
-				} );
-				beforeEachWithConfig( testNotifierConfig );
-
-				componentInstance.notification = testNotification;
-				componentFixture.detectChanges();
-
-				const shiftCallback: () => {} = jasmine.createSpy( 'shiftCallback' );
-				const shiftDistance: number = 100;
-				componentInstance.shift( shiftDistance, false ).then( shiftCallback );
-				tick();
-
-				expect( componentFixture.debugElement.styles[ 'transform' ] )
-					.toBe( `translate3d( 0, ${ shiftDistance + testNotifierConfig.position.vertical.gap }px, 0 )` );
-				expect( shiftCallback ).toHaveBeenCalled();
-
-			} ) );
-
-			it( 'should shift to fill place on bottom (with animations)', fakeAsync( () => {
-
-				// Setup test module
-				const testNotifierConfig: NotifierConfig = new NotifierConfig( {
-					behaviour: {
-						autoHide: false
-					},
-					position: {
-						horizontal: {
-							distance: 12,
-							position: 'left'
-						},
-						vertical: {
-							distance: 12,
-							gap: 10,
-							position: 'bottom'
-						}
-					}
-				} );
-				beforeEachWithConfig( testNotifierConfig );
-
-				componentInstance.notification = testNotification;
-				componentFixture.detectChanges();
-
-				// Mock away the Web Animations API
-				const shiftDistance: number = 100;
-				spyOn( componentFixture.nativeElement, 'animate' ).and.callFake( () => {
-					return {
-						finished: new Promise<undefined>( ( resolve: () => void, reject: () => void ) => {
-							componentFixture.debugElement.styles[ 'transform' ] =
-								`translate3d( 0, ${ shiftDistance + testNotifierConfig.position.vertical.gap }px, 0 )`; // Fake animation result
-							resolve();
-						} )
-					};
-				} );
-
-				const shiftCallback: () => {} = jasmine.createSpy( 'shiftCallback' );
-				componentInstance.shift( shiftDistance, false ).then( shiftCallback );
-				tick();
-
-				expect( componentFixture.debugElement.styles[ 'transform' ] )
-					.toBe( `translate3d( 0, ${ shiftDistance + testNotifierConfig.position.vertical.gap }px, 0 )` );
-				expect( shiftCallback ).toHaveBeenCalled();
-
-			} ) );
-
-			it( 'should shift to make place in the middle', fakeAsync( () => {
-
-				// Setup test module
-				const testNotifierConfig: NotifierConfig = new NotifierConfig( {
-					animations: {
-						enabled: false
-					},
-					behaviour: {
-						autoHide: false
-					},
-					position: {
-						horizontal: {
-							distance: 12,
-							position: 'middle'
-						},
-						vertical: {
-							distance: 12,
-							gap: 10,
-							position: 'top'
-						}
-					}
-				} );
-				beforeEachWithConfig( testNotifierConfig );
-
-				componentInstance.notification = testNotification;
-				componentFixture.detectChanges();
-
-				const shiftCallback: () => {} = jasmine.createSpy( 'shiftCallback' );
-				const shiftDistance: number = 100;
-				componentInstance.shift( shiftDistance, true ).then( shiftCallback );
-				tick();
-
-				expect( componentFixture.debugElement.styles[ 'transform' ] )
-					.toBe( `translate3d( -50%, ${ shiftDistance + testNotifierConfig.position.vertical.gap }px, 0 )` );
-				expect( shiftCallback ).toHaveBeenCalled();
-
-			} ) );
-
-			it( 'should shift to make place in the middle (with animations)', fakeAsync( () => {
-
-				// Setup test module
-				const testNotifierConfig: NotifierConfig = new NotifierConfig( {
-					animations: {
-						enabled: false
-					},
-					behaviour: {
-						autoHide: false
-					},
-					position: {
-						horizontal: {
-							distance: 12,
-							position: 'middle'
-						},
-						vertical: {
-							distance: 12,
-							gap: 10,
-							position: 'top'
-						}
-					}
-				} );
-				beforeEachWithConfig( testNotifierConfig );
-
-				componentInstance.notification = testNotification;
-				componentFixture.detectChanges();
-
-				// Mock away the Web Animations API
-				const shiftDistance: number = 100;
-				spyOn( componentFixture.nativeElement, 'animate' ).and.callFake( () => {
-					return {
-						finished: new Promise<undefined>( ( resolve: () => void, reject: () => void ) => {
-							componentFixture.debugElement.styles[ 'transform' ] =
-								`translate3d( -50%, ${ shiftDistance + testNotifierConfig.position.vertical.gap }px, 0 )`; // Fake animation result
-							resolve();
-						} )
-					};
-				} );
-
-				const shiftCallback: () => {} = jasmine.createSpy( 'shiftCallback' );
-				componentInstance.shift( shiftDistance, true ).then( shiftCallback );
-				tick();
-
-				expect( componentFixture.debugElement.styles[ 'transform' ] )
-					.toBe( `translate3d( -50%, ${ shiftDistance + testNotifierConfig.position.vertical.gap }px, 0 )` );
-				expect( shiftCallback ).toHaveBeenCalled();
-
-			} ) );
+				}
+			} );
+			beforeEachWithConfig( testNotifierConfig );
+
+			componentInstance.notification = testNotification;
+			componentFixture.detectChanges();
+
+			// Check position
+			expect( componentFixture.debugElement.styles[ 'top' ] ).toBe( `${ testNotifierConfig.position.vertical.distance }px` );
 
 		} );
 
-		describe( '(behaviour)', () => {
+		it( 'should render on the bottom', () => {
 
-			it( 'should hide automatically after timeout', fakeAsync( () => {
-
-				// Setup test module
-				beforeEachWithConfig( new NotifierConfig( {
-					animations: {
-						enabled: false
+			// Setup test module
+			const testNotifierConfig: NotifierConfig = new NotifierConfig( {
+				position: {
+					horizontal: {
+						distance: 10,
+						position: 'left'
 					},
-					behaviour: {
-						autoHide: 5000
+					vertical: {
+						distance: 10,
+						gap: 4,
+						position: 'bottom'
 					}
-				} ) );
+				}
+			} );
+			beforeEachWithConfig( testNotifierConfig );
 
-				componentInstance.notification = testNotification;
-				componentFixture.detectChanges();
+			componentInstance.notification = testNotification;
+			componentFixture.detectChanges();
 
-				componentInstance.show();
-				spyOn( componentInstance, 'onClickDismiss' ).and.callThrough();
-				tick();
-
-				timerService.finishManually();
-				tick();
-
-				expect( componentInstance.onClickDismiss ).toHaveBeenCalled();
-
-			} ) );
-
-			it( 'should hide after clicking the dismiss button', fakeAsync( () => {
-
-				// Setup test module
-				beforeEachWithConfig( new NotifierConfig( {
-					animations: {
-						enabled: false
-					},
-					behaviour: {
-						autoHide: false,
-						showDismissButton: true
-					}
-				} ) );
-
-				componentInstance.notification = testNotification;
-				componentFixture.detectChanges();
-
-				componentInstance.show();
-				spyOn( componentInstance, 'onClickDismiss' ).and.callThrough();
-
-				const dismissButtonElement: DebugElement = componentFixture.debugElement.query( By.css( '.x-notifier__notification-button' ) );
-				dismissButtonElement.nativeElement.click(); // Emulate click event
-				componentFixture.detectChanges();
-
-				expect( componentInstance.onClickDismiss ).toHaveBeenCalled();
-
-			} ) );
-
-			it( 'should hide after clicking on the notification', fakeAsync( () => {
-
-				// Setup test module
-				beforeEachWithConfig( new NotifierConfig( {
-					animations: {
-						enabled: false
-					},
-					behaviour: {
-						autoHide: false,
-						onClick: 'hide'
-					}
-				} ) );
-
-				componentInstance.notification = testNotification;
-				componentFixture.detectChanges();
-
-				componentInstance.show();
-				spyOn( componentInstance, 'onClickDismiss' ).and.callThrough();
-
-				componentFixture.nativeElement.click(); // Emulate click event
-				componentFixture.detectChanges();
-
-				expect( componentInstance.onClickDismiss ).toHaveBeenCalled();
-
-			} ) );
-
-			it( 'should not hide after clicking on the notification', fakeAsync( () => {
-
-				// Setup test module
-				beforeEachWithConfig( new NotifierConfig( {
-					animations: {
-						enabled: false
-					},
-					behaviour: {
-						autoHide: false,
-						onClick: false
-					}
-				} ) );
-
-				componentInstance.notification = testNotification;
-				componentFixture.detectChanges();
-
-				componentInstance.show();
-				spyOn( componentInstance, 'onClickDismiss' ).and.callThrough();
-
-				componentFixture.nativeElement.click(); // Emulate click event
-				componentFixture.detectChanges();
-
-				expect( componentInstance.onClickDismiss ).not.toHaveBeenCalled();
-
-			} ) );
-
-			it( 'should pause the autoHide timer on mouseover, and resume again on mouseout', fakeAsync( () => {
-
-				// Setup test module
-				beforeEachWithConfig( new NotifierConfig( {
-					animations: {
-						enabled: false
-					},
-					behaviour: {
-						autoHide: 5000,
-						onMouseover: 'pauseAutoHide'
-					}
-				} ) );
-
-				componentInstance.notification = testNotification;
-				componentFixture.detectChanges();
-
-				componentInstance.show();
-				spyOn( componentInstance, 'onClickDismiss' ).and.callThrough();
-				spyOn( timerService, 'pause' ).and.callThrough();
-				spyOn( timerService, 'continue' ).and.callThrough();
-
-				componentInstance.onNotificationMouseover();
-
-				expect( timerService.pause ).toHaveBeenCalled();
-
-				componentInstance.onNotificationMouseout();
-
-				expect( timerService.continue ).toHaveBeenCalled();
-
-				timerService.finishManually();
-				tick();
-
-				expect( componentInstance.onClickDismiss ).toHaveBeenCalled();
-
-			} ) );
-
-			it( 'should restart the autoHide timer on mouseover', fakeAsync( () => {
-
-				// Setup test module
-				beforeEachWithConfig( new NotifierConfig( {
-					animations: {
-						enabled: false
-					},
-					behaviour: {
-						autoHide: 5000,
-						onMouseover: 'resetAutoHide'
-					}
-				} ) );
-
-				componentInstance.notification = testNotification;
-				componentFixture.detectChanges();
-
-				componentInstance.show();
-				spyOn( componentInstance, 'onClickDismiss' ).and.callThrough();
-				spyOn( timerService, 'stop' ).and.callThrough();
-				spyOn( timerService, 'start' ).and.callThrough();
-
-				componentInstance.onNotificationMouseover();
-
-				expect( timerService.stop ).toHaveBeenCalled();
-
-				componentInstance.onNotificationMouseout();
-
-				expect( timerService.start ).toHaveBeenCalled();
-
-				timerService.finishManually();
-				tick();
-
-				expect( componentInstance.onClickDismiss ).toHaveBeenCalled();
-
-			} ) );
+			// Check position
+			expect( componentFixture.debugElement.styles[ 'bottom' ] ).toBe( `${ testNotifierConfig.position.vertical.distance }px` );
 
 		} );
-
-		/**
-		 * Helper for upfront configuration
-		 */
-		function beforeEachWithConfig( testNotifierConfig: NotifierConfig ): void {
-
-			TestBed
-				.configureTestingModule( {
-					declarations: [
-						NotifierNotificationComponent
-					],
-					providers: [
-						{
-							provide: NotifierService,
-							useValue: {
-								getConfig: () => testNotifierConfig
-							}
-						},
-						{
-							provide: NotifierAnimationService,
-							useClass: MockNotifierAnimationService
-						}
-					]
-				} )
-				.overrideComponent( NotifierNotificationComponent, {
-					set: {
-						providers: [ // Override component-specific providers
-							{
-								provide: NotifierTimerService,
-								useClass: MockNotifierTimerService
-							}
-						]
-					}
-				} );
-			componentFixture = TestBed.createComponent( NotifierNotificationComponent );
-			componentInstance = componentFixture.componentInstance;
-
-			// Get the service from the component's local injector
-			timerService = <MockNotifierTimerService> componentFixture.debugElement.injector.get( NotifierTimerService );
-
-		}
 
 	} );
 
-}
+	describe( '(show)', () => {
+
+		it( 'should show', fakeAsync( () => {
+
+			// Setup test module
+			beforeEachWithConfig( new NotifierConfig( {
+				animations: {
+					enabled: false
+				},
+				behaviour: {
+					autoHide: false
+				}
+			} ) );
+
+			componentInstance.notification = testNotification;
+			componentFixture.detectChanges();
+
+			const showCallback: () => {} = jest.fn();
+			componentInstance.show().then( showCallback );
+			tick();
+
+			expect( componentFixture.debugElement.styles[ 'visibility' ] ).toBe( 'visible' );
+			expect( showCallback ).toHaveBeenCalled();
+
+		} ) );
+
+		it( 'should show (with animations)', fakeAsync( () => {
+
+			// Setup test module
+			beforeEachWithConfig( new NotifierConfig( {
+				behaviour: {
+					autoHide: false
+				}
+			} ) );
+
+			componentInstance.notification = testNotification;
+			componentFixture.detectChanges();
+
+			// Mock away the Web Animations API
+			jest.spyOn( componentFixture.nativeElement, 'animate' ).mockImplementation( () => {
+				return {
+					finished: new Promise<undefined>( ( resolve: () => void, reject: () => void ) => {
+						componentFixture.debugElement.styles[ 'opacity' ] = '1'; // Fake animation result
+						resolve();
+					} )
+				};
+			} );
+
+			const showCallback: () => {} = jest.fn();
+			componentInstance.show().then( showCallback );
+			tick();
+
+			expect( componentFixture.debugElement.styles[ 'visibility' ] ).toBe( 'visible' );
+			expect( componentFixture.debugElement.styles[ 'opacity' ] ).toBe( '1' );
+			expect( showCallback ).toHaveBeenCalled();
+
+		} ) );
+
+	} );
+
+	describe( '(hide)', () => {
+
+		it( 'should hide', fakeAsync( () => {
+
+			// Setup test module
+			beforeEachWithConfig( new NotifierConfig( {
+				animations: {
+					enabled: false
+				},
+				behaviour: {
+					autoHide: false
+				}
+			} ) );
+
+			componentInstance.notification = testNotification;
+			componentFixture.detectChanges();
+
+			const hideCallback: () => {} = jest.fn();
+			componentInstance.hide().then( hideCallback );
+			tick();
+
+			expect( hideCallback ).toHaveBeenCalled();
+
+		} ) );
+
+		it( 'should hide (with animations)', fakeAsync( () => {
+
+			// Setup test module
+			beforeEachWithConfig( new NotifierConfig( {
+				behaviour: {
+					autoHide: false
+				}
+			} ) );
+
+			componentInstance.notification = testNotification;
+			componentFixture.detectChanges();
+
+			// Mock away the Web Animations API
+			jest.spyOn( componentFixture.nativeElement, 'animate' ).mockImplementation( () => {
+				return {
+					finished: new Promise<undefined>( ( resolve: () => void, reject: () => void ) => {
+						componentFixture.debugElement.styles[ 'opacity' ] = '0'; // Fake animation result
+						resolve();
+					} )
+				};
+			} );
+
+			const hideCallback: () => {} = jest.fn();
+			componentInstance.hide().then( hideCallback );
+			tick();
+
+			expect( componentFixture.debugElement.styles[ 'opacity' ] ).toBe( '0' );
+			expect( hideCallback ).toHaveBeenCalled();
+
+		} ) );
+
+	} );
+
+	describe( '(shift)', () => {
+
+		it( 'should shift to make place on top', fakeAsync( () => {
+
+			// Setup test module
+			const testNotifierConfig: NotifierConfig = new NotifierConfig( {
+				animations: {
+					enabled: false
+				},
+				behaviour: {
+					autoHide: false
+				},
+				position: {
+					horizontal: {
+						distance: 12,
+						position: 'left'
+					},
+					vertical: {
+						distance: 12,
+						gap: 10,
+						position: 'top'
+					}
+				}
+			} );
+			beforeEachWithConfig( testNotifierConfig );
+
+			componentInstance.notification = testNotification;
+			componentFixture.detectChanges();
+
+			const shiftCallback: () => {} = jest.fn();
+			const shiftDistance: number = 100;
+			componentInstance.shift( shiftDistance, true ).then( shiftCallback );
+			tick();
+
+			expect( componentFixture.debugElement.styles[ 'transform' ] )
+				.toBe( `translate3d( 0, ${ shiftDistance + testNotifierConfig.position.vertical.gap }px, 0 )` );
+			expect( shiftCallback ).toHaveBeenCalled();
+
+		} ) );
+
+		it( 'should shift to make place on top (with animations)', fakeAsync( () => {
+
+			// Setup test module
+			const testNotifierConfig: NotifierConfig = new NotifierConfig( {
+				behaviour: {
+					autoHide: false
+				},
+				position: {
+					horizontal: {
+						distance: 12,
+						position: 'left'
+					},
+					vertical: {
+						distance: 12,
+						gap: 10,
+						position: 'top'
+					}
+				}
+			} );
+			beforeEachWithConfig( testNotifierConfig );
+
+			componentInstance.notification = testNotification;
+			componentFixture.detectChanges();
+
+			const shiftDistance: number = 100;
+
+			// Mock away the Web Animations API
+			jest.spyOn( componentFixture.nativeElement, 'animate' ).mockImplementation( () => {
+				return {
+					finished: new Promise<undefined>( ( resolve: () => void, reject: () => void ) => {
+						componentFixture.debugElement.styles[ 'transform' ] =
+							`translate3d( 0, ${ shiftDistance + testNotifierConfig.position.vertical.gap }px, 0 )`; // Fake animation result
+						resolve();
+					} )
+				};
+			} );
+
+			const shiftCallback: () => {} = jest.fn();
+			componentInstance.shift( shiftDistance, true ).then( shiftCallback );
+			tick();
+
+			expect( componentFixture.debugElement.styles[ 'transform' ] )
+				.toBe( `translate3d( 0, ${ shiftDistance + testNotifierConfig.position.vertical.gap }px, 0 )` );
+			expect( shiftCallback ).toHaveBeenCalled();
+
+		} ) );
+
+		it( 'should shift to make place on bottom', fakeAsync( () => {
+
+			// Setup test module
+			const testNotifierConfig: NotifierConfig = new NotifierConfig( {
+				animations: {
+					enabled: false
+				},
+				behaviour: {
+					autoHide: false
+				},
+				position: {
+					horizontal: {
+						distance: 12,
+						position: 'left'
+					},
+					vertical: {
+						distance: 12,
+						gap: 10,
+						position: 'bottom'
+					}
+				}
+			} );
+			beforeEachWithConfig( testNotifierConfig );
+
+			componentInstance.notification = testNotification;
+			componentFixture.detectChanges();
+
+			const shiftCallback: () => {} = jest.fn();
+			const shiftDistance: number = 100;
+			componentInstance.shift( shiftDistance, true ).then( shiftCallback );
+			tick();
+
+			expect( componentFixture.debugElement.styles[ 'transform' ] )
+				.toBe( `translate3d( 0, ${ -shiftDistance - testNotifierConfig.position.vertical.gap }px, 0 )` );
+			expect( shiftCallback ).toHaveBeenCalled();
+
+		} ) );
+
+		it( 'should shift to make place on bottom (with animations)', fakeAsync( () => {
+
+			// Setup test module
+			const testNotifierConfig: NotifierConfig = new NotifierConfig( {
+				behaviour: {
+					autoHide: false
+				},
+				position: {
+					horizontal: {
+						distance: 12,
+						position: 'left'
+					},
+					vertical: {
+						distance: 12,
+						gap: 10,
+						position: 'bottom'
+					}
+				}
+			} );
+			beforeEachWithConfig( testNotifierConfig );
+
+			componentInstance.notification = testNotification;
+			componentFixture.detectChanges();
+
+			// Mock away the Web Animations API
+			const shiftDistance: number = 100;
+			jest.spyOn( componentFixture.nativeElement, 'animate' ).mockImplementation( () => {
+				return {
+					finished: new Promise<undefined>( ( resolve: () => void, reject: () => void ) => {
+						componentFixture.debugElement.styles[ 'transform' ] =
+							`translate3d( 0, ${ -shiftDistance - testNotifierConfig.position.vertical.gap }px, 0 )`; // Fake animation result
+						resolve();
+					} )
+				};
+			} );
+
+			const shiftCallback: () => {} = jest.fn();
+			componentInstance.shift( shiftDistance, true ).then( shiftCallback );
+			tick();
+
+			expect( componentFixture.debugElement.styles[ 'transform' ] )
+				.toBe( `translate3d( 0, ${ -shiftDistance - testNotifierConfig.position.vertical.gap }px, 0 )` );
+			expect( shiftCallback ).toHaveBeenCalled();
+
+		} ) );
+
+		it( 'should shift to fill place on top', fakeAsync( () => {
+
+			// Setup test module
+			const testNotifierConfig: NotifierConfig = new NotifierConfig( {
+				animations: {
+					enabled: false
+				},
+				behaviour: {
+					autoHide: false
+				},
+				position: {
+					horizontal: {
+						distance: 12,
+						position: 'left'
+					},
+					vertical: {
+						distance: 12,
+						gap: 10,
+						position: 'top'
+					}
+				}
+			} );
+			beforeEachWithConfig( testNotifierConfig );
+
+			componentInstance.notification = testNotification;
+			componentFixture.detectChanges();
+
+			const shiftCallback: () => {} = jest.fn();
+			const shiftDistance: number = 100;
+			componentInstance.shift( shiftDistance, false ).then( shiftCallback );
+			tick();
+
+			expect( componentFixture.debugElement.styles[ 'transform' ] )
+				.toBe( `translate3d( 0, ${ -shiftDistance - testNotifierConfig.position.vertical.gap }px, 0 )` );
+			expect( shiftCallback ).toHaveBeenCalled();
+
+		} ) );
+
+		it( 'should shift to fill place on top (with animations)', fakeAsync( () => {
+
+			// Setup test module
+			const testNotifierConfig: NotifierConfig = new NotifierConfig( {
+				behaviour: {
+					autoHide: false
+				},
+				position: {
+					horizontal: {
+						distance: 12,
+						position: 'left'
+					},
+					vertical: {
+						distance: 12,
+						gap: 10,
+						position: 'top'
+					}
+				}
+			} );
+			beforeEachWithConfig( testNotifierConfig );
+
+			componentInstance.notification = testNotification;
+			componentFixture.detectChanges();
+
+			// Mock away the Web Animations API
+			const shiftDistance: number = 100;
+			jest.spyOn( componentFixture.nativeElement, 'animate' ).mockImplementation( () => {
+				return {
+					finished: new Promise<undefined>( ( resolve: () => void, reject: () => void ) => {
+						componentFixture.debugElement.styles[ 'transform' ] =
+							`translate3d( 0, ${ -shiftDistance - testNotifierConfig.position.vertical.gap }px, 0 )`; // Fake animation result
+						resolve();
+					} )
+				};
+			} );
+
+			const shiftCallback: () => {} = jest.fn();
+			componentInstance.shift( shiftDistance, false ).then( shiftCallback );
+			tick();
+
+			expect( componentFixture.debugElement.styles[ 'transform' ] )
+				.toBe( `translate3d( 0, ${ 0 - shiftDistance - testNotifierConfig.position.vertical.gap }px, 0 )` );
+			expect( shiftCallback ).toHaveBeenCalled();
+
+		} ) );
+
+		it( 'should shift to fill place on bottom', fakeAsync( () => {
+
+			// Setup test module
+			const testNotifierConfig: NotifierConfig = new NotifierConfig( {
+				animations: {
+					enabled: false
+				},
+				behaviour: {
+					autoHide: false
+				},
+				position: {
+					horizontal: {
+						distance: 12,
+						position: 'left'
+					},
+					vertical: {
+						distance: 12,
+						gap: 10,
+						position: 'bottom'
+					}
+				}
+			} );
+			beforeEachWithConfig( testNotifierConfig );
+
+			componentInstance.notification = testNotification;
+			componentFixture.detectChanges();
+
+			const shiftCallback: () => {} = jest.fn();
+			const shiftDistance: number = 100;
+			componentInstance.shift( shiftDistance, false ).then( shiftCallback );
+			tick();
+
+			expect( componentFixture.debugElement.styles[ 'transform' ] )
+				.toBe( `translate3d( 0, ${ shiftDistance + testNotifierConfig.position.vertical.gap }px, 0 )` );
+			expect( shiftCallback ).toHaveBeenCalled();
+
+		} ) );
+
+		it( 'should shift to fill place on bottom (with animations)', fakeAsync( () => {
+
+			// Setup test module
+			const testNotifierConfig: NotifierConfig = new NotifierConfig( {
+				behaviour: {
+					autoHide: false
+				},
+				position: {
+					horizontal: {
+						distance: 12,
+						position: 'left'
+					},
+					vertical: {
+						distance: 12,
+						gap: 10,
+						position: 'bottom'
+					}
+				}
+			} );
+			beforeEachWithConfig( testNotifierConfig );
+
+			componentInstance.notification = testNotification;
+			componentFixture.detectChanges();
+
+			// Mock away the Web Animations API
+			const shiftDistance: number = 100;
+			jest.spyOn( componentFixture.nativeElement, 'animate' ).mockImplementation( () => {
+				return {
+					finished: new Promise<undefined>( ( resolve: () => void, reject: () => void ) => {
+						componentFixture.debugElement.styles[ 'transform' ] =
+							`translate3d( 0, ${ shiftDistance + testNotifierConfig.position.vertical.gap }px, 0 )`; // Fake animation result
+						resolve();
+					} )
+				};
+			} );
+
+			const shiftCallback: () => {} = jest.fn();
+			componentInstance.shift( shiftDistance, false ).then( shiftCallback );
+			tick();
+
+			expect( componentFixture.debugElement.styles[ 'transform' ] )
+				.toBe( `translate3d( 0, ${ shiftDistance + testNotifierConfig.position.vertical.gap }px, 0 )` );
+			expect( shiftCallback ).toHaveBeenCalled();
+
+		} ) );
+
+		it( 'should shift to make place in the middle', fakeAsync( () => {
+
+			// Setup test module
+			const testNotifierConfig: NotifierConfig = new NotifierConfig( {
+				animations: {
+					enabled: false
+				},
+				behaviour: {
+					autoHide: false
+				},
+				position: {
+					horizontal: {
+						distance: 12,
+						position: 'middle'
+					},
+					vertical: {
+						distance: 12,
+						gap: 10,
+						position: 'top'
+					}
+				}
+			} );
+			beforeEachWithConfig( testNotifierConfig );
+
+			componentInstance.notification = testNotification;
+			componentFixture.detectChanges();
+
+			const shiftCallback: () => {} = jest.fn();
+			const shiftDistance: number = 100;
+			componentInstance.shift( shiftDistance, true ).then( shiftCallback );
+			tick();
+
+			expect( componentFixture.debugElement.styles[ 'transform' ] )
+				.toBe( `translate3d( -50%, ${ shiftDistance + testNotifierConfig.position.vertical.gap }px, 0 )` );
+			expect( shiftCallback ).toHaveBeenCalled();
+
+		} ) );
+
+		it( 'should shift to make place in the middle (with animations)', fakeAsync( () => {
+
+			// Setup test module
+			const testNotifierConfig: NotifierConfig = new NotifierConfig( {
+				animations: {
+					enabled: false
+				},
+				behaviour: {
+					autoHide: false
+				},
+				position: {
+					horizontal: {
+						distance: 12,
+						position: 'middle'
+					},
+					vertical: {
+						distance: 12,
+						gap: 10,
+						position: 'top'
+					}
+				}
+			} );
+			beforeEachWithConfig( testNotifierConfig );
+
+			componentInstance.notification = testNotification;
+			componentFixture.detectChanges();
+
+			// Mock away the Web Animations API
+			const shiftDistance: number = 100;
+			jest.spyOn( componentFixture.nativeElement, 'animate' ).mockImplementation( () => {
+				return {
+					finished: new Promise<undefined>( ( resolve: () => void, reject: () => void ) => {
+						componentFixture.debugElement.styles[ 'transform' ] =
+							`translate3d( -50%, ${ shiftDistance + testNotifierConfig.position.vertical.gap }px, 0 )`; // Fake animation result
+						resolve();
+					} )
+				};
+			} );
+
+			const shiftCallback: () => {} = jest.fn();
+			componentInstance.shift( shiftDistance, true ).then( shiftCallback );
+			tick();
+
+			expect( componentFixture.debugElement.styles[ 'transform' ] )
+				.toBe( `translate3d( -50%, ${ shiftDistance + testNotifierConfig.position.vertical.gap }px, 0 )` );
+			expect( shiftCallback ).toHaveBeenCalled();
+
+		} ) );
+
+	} );
+
+	describe( '(behaviour)', () => {
+
+		it( 'should hide automatically after timeout', fakeAsync( () => {
+
+			// Setup test module
+			beforeEachWithConfig( new NotifierConfig( {
+				animations: {
+					enabled: false
+				},
+				behaviour: {
+					autoHide: 5000
+				}
+			} ) );
+
+			componentInstance.notification = testNotification;
+			componentFixture.detectChanges();
+
+			componentInstance.show();
+			jest.spyOn( componentInstance, 'onClickDismiss' );
+			tick();
+
+			timerService.finishManually();
+			tick();
+
+			expect( componentInstance.onClickDismiss ).toHaveBeenCalled();
+
+		} ) );
+
+		it( 'should hide after clicking the dismiss button', fakeAsync( () => {
+
+			// Setup test module
+			beforeEachWithConfig( new NotifierConfig( {
+				animations: {
+					enabled: false
+				},
+				behaviour: {
+					autoHide: false,
+					showDismissButton: true
+				}
+			} ) );
+
+			componentInstance.notification = testNotification;
+			componentFixture.detectChanges();
+
+			componentInstance.show();
+			jest.spyOn( componentInstance, 'onClickDismiss' );
+
+			const dismissButtonElement: DebugElement = componentFixture.debugElement.query( By.css( '.x-notifier__notification-button' ) );
+			dismissButtonElement.nativeElement.click(); // Emulate click event
+			componentFixture.detectChanges();
+
+			expect( componentInstance.onClickDismiss ).toHaveBeenCalled();
+
+		} ) );
+
+		it( 'should hide after clicking on the notification', fakeAsync( () => {
+
+			// Setup test module
+			beforeEachWithConfig( new NotifierConfig( {
+				animations: {
+					enabled: false
+				},
+				behaviour: {
+					autoHide: false,
+					onClick: 'hide'
+				}
+			} ) );
+
+			componentInstance.notification = testNotification;
+			componentFixture.detectChanges();
+
+			componentInstance.show();
+			jest.spyOn( componentInstance, 'onClickDismiss' );
+
+			componentFixture.nativeElement.click(); // Emulate click event
+			componentFixture.detectChanges();
+
+			expect( componentInstance.onClickDismiss ).toHaveBeenCalled();
+
+		} ) );
+
+		it( 'should not hide after clicking on the notification', fakeAsync( () => {
+
+			// Setup test module
+			beforeEachWithConfig( new NotifierConfig( {
+				animations: {
+					enabled: false
+				},
+				behaviour: {
+					autoHide: false,
+					onClick: false
+				}
+			} ) );
+
+			componentInstance.notification = testNotification;
+			componentFixture.detectChanges();
+
+			componentInstance.show();
+			jest.spyOn( componentInstance, 'onClickDismiss' );
+
+			componentFixture.nativeElement.click(); // Emulate click event
+			componentFixture.detectChanges();
+
+			expect( componentInstance.onClickDismiss ).not.toHaveBeenCalled();
+
+		} ) );
+
+		it( 'should pause the autoHide timer on mouseover, and resume again on mouseout', fakeAsync( () => {
+
+			// Setup test module
+			beforeEachWithConfig( new NotifierConfig( {
+				animations: {
+					enabled: false
+				},
+				behaviour: {
+					autoHide: 5000,
+					onMouseover: 'pauseAutoHide'
+				}
+			} ) );
+
+			componentInstance.notification = testNotification;
+			componentFixture.detectChanges();
+
+			componentInstance.show();
+			jest.spyOn( componentInstance, 'onClickDismiss' );
+			jest.spyOn( timerService, 'pause' );
+			jest.spyOn( timerService, 'continue' );
+
+			componentInstance.onNotificationMouseover();
+
+			expect( timerService.pause ).toHaveBeenCalled();
+
+			componentInstance.onNotificationMouseout();
+
+			expect( timerService.continue ).toHaveBeenCalled();
+
+			timerService.finishManually();
+			tick();
+
+			expect( componentInstance.onClickDismiss ).toHaveBeenCalled();
+
+		} ) );
+
+		it( 'should restart the autoHide timer on mouseover', fakeAsync( () => {
+
+			// Setup test module
+			beforeEachWithConfig( new NotifierConfig( {
+				animations: {
+					enabled: false
+				},
+				behaviour: {
+					autoHide: 5000,
+					onMouseover: 'resetAutoHide'
+				}
+			} ) );
+
+			componentInstance.notification = testNotification;
+			componentFixture.detectChanges();
+
+			componentInstance.show();
+			jest.spyOn( componentInstance, 'onClickDismiss' );
+			jest.spyOn( timerService, 'stop' );
+			jest.spyOn( timerService, 'start' );
+
+			componentInstance.onNotificationMouseover();
+
+			expect( timerService.stop ).toHaveBeenCalled();
+
+			componentInstance.onNotificationMouseout();
+
+			expect( timerService.start ).toHaveBeenCalled();
+
+			timerService.finishManually();
+			tick();
+
+			expect( componentInstance.onClickDismiss ).toHaveBeenCalled();
+
+		} ) );
+
+	} );
+
+	/**
+	 * Helper for upfront configuration
+	 */
+	function beforeEachWithConfig( testNotifierConfig: NotifierConfig ): void {
+
+		TestBed
+			.configureTestingModule( {
+				declarations: [
+					NotifierNotificationComponent
+				],
+				providers: [
+					{
+						provide: NotifierService,
+						useValue: {
+							getConfig: () => testNotifierConfig
+						}
+					},
+					{
+						provide: NotifierAnimationService,
+						useClass: MockNotifierAnimationService
+					}
+				]
+			} )
+			.overrideComponent( NotifierNotificationComponent, {
+				set: {
+					providers: [ // Override component-specific providers
+						{
+							provide: NotifierTimerService,
+							useClass: MockNotifierTimerService
+						}
+					]
+				}
+			} );
+		componentFixture = TestBed.createComponent( NotifierNotificationComponent );
+		componentInstance = componentFixture.componentInstance;
+
+		// Get the service from the component's local injector
+		timerService = <MockNotifierTimerService> componentFixture.debugElement.injector.get( NotifierTimerService );
+
+	}
+
+} );
 
 /**
  * Mock notifier animation service, always returning the animation
