+++ src/lib/src/services/notifier-timer.service.spec.ts
@@ -5,93 +5,91 @@ import { NotifierTimerService } from './notifier-timer.service';
 /**
  * Notifier Timer Service - Unit Test
  */
-export function main(): void {
+describe( 'Notifier Timer Service', () => {
 
-	describe( 'Notifier Timer Service', () => {
+	const fullAnimationTime: number = 5000;
+	const longAnimationTime: number = 4000;
+	const shortAnimationTime: number = 1000;
 
-		const fullAnimationTime: number = 5000;
-		const longAnimationTime: number = 4000;
-		const shortAnimationTime: number = 1000;
+	let timerService: NotifierTimerService;
+	let mockDate: MockDate;
 
-		let timerService: NotifierTimerService;
-		let mockDate: MockDate = new MockDate();
-
-		// Setup test module
-		beforeEach( () => {
-			TestBed.configureTestingModule( {
-				providers: [
-					NotifierTimerService
-				]
-			} );
+	// Setup test module
+	beforeEach( () => {
+		TestBed.configureTestingModule( {
+			providers: [
+				NotifierTimerService
+			]
 		} );
+	} );
 
-		// Inject dependencies
-		beforeEach( inject( [ NotifierTimerService ], ( notifierTimerService: NotifierTimerService ) => {
-			timerService = notifierTimerService;
-			mockDate.reset(); // Reset global date
-		} ) );
+	// Inject dependencies
+	beforeEach( inject( [ NotifierTimerService ], ( notifierTimerService: NotifierTimerService ) => {
+		timerService = notifierTimerService;
+		mockDate = new MockDate();
+	} ) );
 
-		it( 'should instantiate', () => {
+	it( 'should instantiate', () => {
 
-			expect( timerService ).toBeDefined();
-
-		} );
-
-		it( 'should start and stop the timer', fakeAsync( () => {
-
-			const timerServiceCallback: () => {} = jasmine.createSpy( 'timerServiceCallback' );
-			timerService.start( fullAnimationTime ).then( timerServiceCallback );
-
-			tick( longAnimationTime );
-
-			expect( timerServiceCallback ).not.toHaveBeenCalled();
-
-			tick( shortAnimationTime );
-
-			expect( timerServiceCallback ).toHaveBeenCalled();
-
-		} ) );
-
-		it( 'should pause and resume the timer', fakeAsync( () => {
-
-			spyOn( window, 'Date' ).and.callFake( () => mockDate );
-			const timerServiceCallback: () => {} = jasmine.createSpy( 'timerServiceCallback' );
-			timerService.start( fullAnimationTime ).then( timerServiceCallback );
-
-			tick( longAnimationTime );
-			mockDate.fastForwardTime( longAnimationTime ); // Also update the global Date (in addition to the tick)
-
-			timerService.pause();
-
-			tick( shortAnimationTime );
-			mockDate.fastForwardTime( shortAnimationTime ); // Also update the global Date (in addition to the tick)
-
-			expect( timerServiceCallback ).not.toHaveBeenCalled();
-
-			// Resumes the timer, using the same duration as above (a continue doesn't exist yet)
-			timerService.continue();
-			tick( shortAnimationTime );
-
-			expect( timerServiceCallback ).toHaveBeenCalled();
-
-		} ) );
-
-		it( 'should stop the timer', fakeAsync( () => {
-
-			const timerServiceCallback: () => {} = jasmine.createSpy( 'timerServiceCallback' );
-			timerService.start( fullAnimationTime ).then( timerServiceCallback );
-
-			tick( longAnimationTime );
-			timerService.stop();
-			tick( shortAnimationTime );
-
-			expect( timerServiceCallback ).not.toHaveBeenCalled();
-
-		} ) );
+		expect( timerService ).toBeDefined();
 
 	} );
 
-}
+	it( 'should start and stop the timer', fakeAsync( () => {
+
+		const timerServiceCallback: () => {} = jest.fn();
+		timerService.start( fullAnimationTime ).then( timerServiceCallback );
+
+		tick( longAnimationTime );
+
+		expect( timerServiceCallback ).not.toHaveBeenCalled();
+
+		tick( shortAnimationTime );
+
+		expect( timerServiceCallback ).toHaveBeenCalled();
+
+	} ) );
+
+	it( 'should pause and resume the timer', fakeAsync( () => {
+
+		// tslint:disable no-any
+		jest.spyOn( <any> window, 'Date' ).mockImplementation( () => mockDate );
+		// tslint:enable no-any
+		const timerServiceCallback: () => {} = jest.fn();
+		timerService.start( fullAnimationTime ).then( timerServiceCallback );
+
+		tick( longAnimationTime );
+		mockDate.fastForwardTime( longAnimationTime ); // Also update the global Date (in addition to the tick)
+
+		timerService.pause();
+
+		tick( shortAnimationTime );
+		mockDate.fastForwardTime( shortAnimationTime ); // Also update the global Date (in addition to the tick)
+
+		expect( timerServiceCallback ).not.toHaveBeenCalled();
+
+		// Resumes the timer, using the same duration as above (a continue doesn't exist yet)
+		timerService.continue();
+		tick( shortAnimationTime );
+
+		expect( timerServiceCallback ).toHaveBeenCalled();
+
+	} ) );
+
+	it( 'should stop the timer', fakeAsync( () => {
+
+		const timerServiceCallback: () => {} = jest.fn();
+		timerService.start( fullAnimationTime ).then( timerServiceCallback );
+
+		tick( longAnimationTime );
+		timerService.stop();
+		tick( shortAnimationTime );
+
+		expect( timerServiceCallback ).not.toHaveBeenCalled();
+
+	} ) );
+
+} );
 
 /**
  * Mock Date, allows for fast-forwarding the time even in the global Date object
@@ -107,14 +105,6 @@ class MockDate extends Date {
 	 * Elapsed time (since init)
 	 */
 	private elapsedTime: number;
-
-	/**
-	 * Reset the date to the default values (thus, the current time as well)
-	 */
-	public reset(): void {
-		this.startTime = new Date().getTime();
-		this.elapsedTime = 0;
-	}
 
 	/**
 	 * Fast-forward the current time manually
