+++ src/lib/src/services/notifier.service.spec.ts
@@ -9,190 +9,186 @@ import { NotifierService } from './notifier.service';
 /**
  * Notifier Service - Unit Test
  */
-export function main(): void {
-
-	const testNotifierConfig: NotifierConfig = {
-		animations: {
-			enabled: true,
-			hide: {
-				easing: 'ease',
-				offset: 50,
-				preset: 'fade',
-				speed: 300
-			},
-			overlap: 150,
-			shift: {
-				easing: 'ease',
-				speed: 300
-			},
-			show: {
-				easing: 'ease',
-				preset: 'slide',
-				speed: 300
-			}
+const testNotifierConfig: NotifierConfig = {
+	animations: {
+		enabled: true,
+		hide: {
+			easing: 'ease',
+			offset: 50,
+			preset: 'fade',
+			speed: 300
 		},
-		behaviour: {
-			autoHide: 7000,
-			onClick: false,
-			onMouseover: 'pauseAutoHide',
-			showDismissButton: true,
-			stacking: 4
+		overlap: 150,
+		shift: {
+			easing: 'ease',
+			speed: 300
 		},
-		position: {
-			horizontal: {
-				distance: 12,
-				position: 'left'
-			},
-			vertical: {
-				distance: 12,
-				gap: 10,
-				position: 'bottom'
-			}
+		show: {
+			easing: 'ease',
+			preset: 'slide',
+			speed: 300
+		}
+	},
+	behaviour: {
+		autoHide: 7000,
+		onClick: false,
+		onMouseover: 'pauseAutoHide',
+		showDismissButton: true,
+		stacking: 4
+	},
+	position: {
+		horizontal: {
+			distance: 12,
+			position: 'left'
 		},
-		theme: 'material'
-	};
+		vertical: {
+			distance: 12,
+			gap: 10,
+			position: 'bottom'
+		}
+	},
+	theme: 'material'
+};
 
-	describe( 'Notifier Service', () => {
+describe( 'Notifier Service', () => {
 
-		let service: NotifierService;
-		let queueService: MockNotifierQueueService;
+	let service: NotifierService;
+	let queueService: MockNotifierQueueService;
 
-		// Setup test module
-		beforeEach( () => {
-			TestBed.configureTestingModule( {
-				providers: [
-					NotifierService,
-					{
-						provide: NotifierQueueService,
-						useClass: MockNotifierQueueService
-					},
-					{
-						provide: NotifierConfig,
-						useValue: testNotifierConfig
-					}
-				]
-			} );
+	// Setup test module
+	beforeEach( () => {
+		TestBed.configureTestingModule( {
+			providers: [
+				NotifierService,
+				{
+					provide: NotifierQueueService,
+					useClass: MockNotifierQueueService
+				},
+				{
+					provide: NotifierConfig,
+					useValue: testNotifierConfig
+				}
+			]
 		} );
+	} );
 
-		// Inject dependencies
-		beforeEach( inject( [ NotifierService, NotifierQueueService ],
-			( notifierService: NotifierService, notifierQueueService: MockNotifierQueueService ) => {
-			service = notifierService;
-			queueService = notifierQueueService;
-			}
-		) );
+	// Inject dependencies
+	beforeEach( inject( [ NotifierService, NotifierQueueService ],
+		( notifierService: NotifierService, notifierQueueService: MockNotifierQueueService ) => {
+		service = notifierService;
+		queueService = notifierQueueService;
+		}
+	) );
 
-		it( 'should instantiate', () => {
+	it( 'should instantiate', () => {
 
-			expect( service ).toBeDefined();
-
-		} );
-
-		it( 'should show a notification', () => {
-
-			const testNotificationOptions: NotifierNotificationOptions = {
-				id: 'ID_FAKE',
-				message: 'Lorem ipsum dolor sit amet.',
-				type: 'SUCCESS'
-			};
-			const expectedAction: NotifierAction = {
-				payload: testNotificationOptions,
-				type: 'SHOW'
-			};
-			service.show( testNotificationOptions );
-
-			expect( queueService.lastAction ).toEqual( expectedAction );
-
-		} );
-
-		it( 'should show a notification, the simply way', () => {
-
-			const testNotificationOptions: NotifierNotificationOptions = {
-				message: 'Lorem ipsum dolor sit amet.',
-				type: 'SUCCESS'
-			};
-			const expectedAction: NotifierAction = {
-				payload: testNotificationOptions,
-				type: 'SHOW'
-			};
-			service.notify( testNotificationOptions.type, testNotificationOptions.message );
-
-			expect( queueService.lastAction ).toEqual( expectedAction );
-
-		} );
-
-		it( 'should show a notification with an explicit ID, the simply way', () => {
-
-			const testNotificationOptions: NotifierNotificationOptions = {
-				id: 'ID_FAKE',
-				message: 'Lorem ipsum dolor sit amet.',
-				type: 'SUCCESS'
-			};
-			const expectedAction: NotifierAction = {
-				payload: testNotificationOptions,
-				type: 'SHOW'
-			};
-			service.notify( testNotificationOptions.type, testNotificationOptions.message, testNotificationOptions.id );
-
-			expect( queueService.lastAction ).toEqual( expectedAction );
-
-		} );
-
-		it( 'should hide a specific notification', () => {
-
-			const testNotificationId: string = 'ID_FAKE';
-			const expectedAction: NotifierAction = {
-				payload: testNotificationId,
-				type: 'HIDE'
-			};
-			service.hide( testNotificationId );
-
-			expect( queueService.lastAction ).toEqual( expectedAction );
-
-		} );
-
-		it( 'should hide the newest notification', () => {
-
-			const expectedAction: NotifierAction = {
-				type: 'HIDE_NEWEST'
-			};
-			service.hideNewest();
-
-			expect( queueService.lastAction ).toEqual( expectedAction );
-
-		} );
-
-		it( 'should hide the oldest notification', () => {
-
-			const expectedAction: NotifierAction = {
-				type: 'HIDE_OLDEST'
-			};
-			service.hideOldest();
-
-			expect( queueService.lastAction ).toEqual( expectedAction );
-
-		} );
-
-		it( 'should hide all notifications', () => {
-
-			const expectedAction: NotifierAction = {
-				type: 'HIDE_ALL'
-			};
-			service.hideAll();
-
-			expect( queueService.lastAction ).toEqual( expectedAction );
-
-		} );
-
-		it( 'should return the configuration', () => {
-
-			expect( service.getConfig() ).toEqual( testNotifierConfig );
-
-		} );
+		expect( service ).toBeDefined();
 
 	} );
 
-}
+	it( 'should show a notification', () => {
+
+		const testNotificationOptions: NotifierNotificationOptions = {
+			id: 'ID_FAKE',
+			message: 'Lorem ipsum dolor sit amet.',
+			type: 'SUCCESS'
+		};
+		const expectedAction: NotifierAction = {
+			payload: testNotificationOptions,
+			type: 'SHOW'
+		};
+		service.show( testNotificationOptions );
+
+		expect( queueService.lastAction ).toEqual( expectedAction );
+
+	} );
+
+	it( 'should show a notification, the simply way', () => {
+
+		const testNotificationOptions: NotifierNotificationOptions = {
+			message: 'Lorem ipsum dolor sit amet.',
+			type: 'SUCCESS'
+		};
+		const expectedAction: NotifierAction = {
+			payload: testNotificationOptions,
+			type: 'SHOW'
+		};
+		service.notify( testNotificationOptions.type, testNotificationOptions.message );
+
+		expect( queueService.lastAction ).toEqual( expectedAction );
+
+	} );
+
+	it( 'should show a notification with an explicit ID, the simply way', () => {
+
+		const testNotificationOptions: NotifierNotificationOptions = {
+			id: 'ID_FAKE',
+			message: 'Lorem ipsum dolor sit amet.',
+			type: 'SUCCESS'
+		};
+		const expectedAction: NotifierAction = {
+			payload: testNotificationOptions,
+			type: 'SHOW'
+		};
+		service.notify( testNotificationOptions.type, testNotificationOptions.message, testNotificationOptions.id );
+
+		expect( queueService.lastAction ).toEqual( expectedAction );
+
+	} );
+
+	it( 'should hide a specific notification', () => {
+
+		const testNotificationId: string = 'ID_FAKE';
+		const expectedAction: NotifierAction = {
+			payload: testNotificationId,
+			type: 'HIDE'
+		};
+		service.hide( testNotificationId );
+
+		expect( queueService.lastAction ).toEqual( expectedAction );
+
+	} );
+
+	it( 'should hide the newest notification', () => {
+
+		const expectedAction: NotifierAction = {
+			type: 'HIDE_NEWEST'
+		};
+		service.hideNewest();
+
+		expect( queueService.lastAction ).toEqual( expectedAction );
+
+	} );
+
+	it( 'should hide the oldest notification', () => {
+
+		const expectedAction: NotifierAction = {
+			type: 'HIDE_OLDEST'
+		};
+		service.hideOldest();
+
+		expect( queueService.lastAction ).toEqual( expectedAction );
+
+	} );
+
+	it( 'should hide all notifications', () => {
+
+		const expectedAction: NotifierAction = {
+			type: 'HIDE_ALL'
+		};
+		service.hideAll();
+
+		expect( queueService.lastAction ).toEqual( expectedAction );
+
+	} );
+
+	it( 'should return the configuration', () => {
+
+		expect( service.getConfig() ).toEqual( testNotifierConfig );
+
+	} );
+
+} );
 
 /**
  * Mock Notifier Queue Service
