+++ tools/jest/jest.config.json
@@ -1,0 +1,10 @@
+{
+	"globals": {
+		"__TS_CONFIG__": "<rootDir>/tools/tsc/tsconfig.spec.json",
+		"__TRANSFORM_HTML__": true
+	},
+	"collectCoverage": true,
+	"coverageDirectory": "<rootDir>/coverage",
+	"preset": "jest-preset-angular",
+	"setupTestFrameworkScriptFile": "<rootDir>/tools/jest/jest.setup.ts"
+}
