+++ tools/tsc/tsconfig.spec.json
@@ -1,0 +1,16 @@
+{
+	"extends": "./../../tsconfig.json",
+	"compilerOptions": {
+		"outDir": "./../../build/tsc/spec",
+		"module": "commonjs",
+		"target": "es5",
+		"baseUrl": "",
+		"types": [
+			"node"
+		]
+	},
+	"include": [
+		"./../../src/lib/**/*.spec.ts",
+		"./../../src/lib/**/*.d.ts"
+	]
+}
