--- tsconfig-aot.json
@@ -1,31 +1,0 @@
-{
-	"compilerOptions": {
-		"declaration": true,
-		"emitDecoratorMetadata": true,
-		"experimentalDecorators": true,
-		"lib": [
-			"dom",
-			"es2015"
-		],
-		"module": "es2015",
-		"moduleResolution": "node",
-		"newLine": "CRLF",
-		"noImplicitAny": true,
-		"noImplicitThis": true,
-		"removeComments": false,
-		"skipLibCheck": true,
-		"sourceMap": true,
-		"strictNullChecks": true,
-		"target": "es5",
-		"typeRoots": [
-			"./node_modules/@types/"
-		]
-	},
-	"files": [
-		"./temp/index.ts"
-	],
-	"angularCompilerOptions": {
-		"skipMetadataEmit": false,
-		"skipTemplateCodegen": true
-	}
-}
