--- tsconfig.dev.json
@@ -1,14 +1,0 @@
-{
-	"extends": "./tsconfig.base.json",
-	"compilerOptions": {
-		"outDir": "./build/tsc/app",
-		"baseUrl": "",
-		"types": [
-			"node"
-		]
-	},
-	"exclude": [
-		"node_modules",
-		"src/**/*.spec.ts"
-	]
-}
